Rework PerformanceTracker to make it usable with WebRTC
Previously PerformanceTracker was expected to be notified at the
moment when each frame is decoded and rendered. It also had
assumption that latest_event_timestamp is known for each frame
before that frame is rendered. That doesn't work with WebRTC
where frame stats will have to be sent separately from each
frame. With this change PerformanceTracker gets FrameStats
struct for each frame. That struct contains all timestamps for
the frame, which allows to log all relevant stats.

BUG=621691

Review-Url: https://codereview.chromium.org/2096643003
Cr-Commit-Position: refs/heads/master@{#402054}