[css-flexbox] Correctly implement freezing of inflexible items
We did not have a correct implementation of steps 2 and 3 of the
"9.7 Resolving flexible lengths" section of the spec:
https://drafts.csswg.org/css-flexbox/#resolve-flexible-lengths

In particular inflexible items (flex factor 0 or flex basis
violating the min/max size constraint) reduce the initial free
space.

This is an issue when sizing flex items with a flex factor of 0..1.

BUG=617956
TEST=css3/flexbox/max-width-violation.html

Review-Url: https://codereview.chromium.org/2096943004
Cr-Commit-Position: refs/heads/master@{#402279}