Add new interface to expose ClientSession details to Host Extensions
Prior to the introduction of HostExtensions, code running in the ClientSession
had access to private state information which it could use to make decisions.
Once this code was moved into a HostExtension, it needed to interact with the
session via the ClientSessionControl interface.  This interface was already
being used for other purposes though so it is too easy to pollute it with
methods/properties which the other consumers of the interface don't care about.

My change adds a new, HostExtension specific interface which wraps the
ClientSessionControl interface.  It also provides a location to add new
methods/properties specifcally for HostExtensions.

This change uses this new interface by adding the session Id property and
exposing it to Host Extensions.  A future CL will update the existing
extensions to ue it.

BUG=591746

Review-Url: https://codereview.chromium.org/2091553002
Cr-Commit-Position: refs/heads/master@{#402288}