[TTS] Update Tap counters to compensate for Quick Answers.
When a Tap causes a Quick Answer that shows a caption with doesAnswer,
the user has no real need to open the panel.

This change counts Answers since the last open and uses them to compensate
for Taps without any panel open so we don't stop resolving and preloading
(which would cause subsequent Quick Answers to not appear automatically).

BUG=622160,609924

Review-Url: https://codereview.chromium.org/2099753003
Cr-Commit-Position: refs/heads/master@{#402351}