Use the "aggressive" V8 code cache strategy for CacheStorage as default.
According to the result of the experiments, "aggressive" seems to be the best strategy.
https://goo.gl/ow2zEy
So now we change the default behavior to "aggressive".

This cl adds the browser test for the default behavior in service_worker_browsertest.cc.
And also this cl changes testing/variations/*.json files. It is because in Chromium
builds, FieldTrialList::FindFullName() returns the first group_name in these files.

BUG=581613

Review-Url: https://codereview.chromium.org/2082253002
Cr-Commit-Position: refs/heads/master@{#402403}