PlzNavigate: Handle navigation within MHTML documents
Within MHTML documents, the content ID scheme ("cid:") is used to specify
navigations to another part of the document. Blink understands how to load
such resources (cf. MHTMLArchive.cpp and ResourceFetcher.cpp), so //content
doesn't need to anything special for them. In particular, it does not need
to make a network request when navigating to such URLs. This CL adds that
behavior so that PlzNavigate correctly handles such navigations.

BUG=623553

Review-Url: https://codereview.chromium.org/2098543003
Cr-Commit-Position: refs/heads/master@{#402438}