Don't commit the blocked URL when a frame is blocked by XFrameOptions.
Previously, when a load was blocked by XFO or frame-ancestors, we
committed a blank page and left the original URL as the committed URL.
In some cases, this led to the browser process thinking that the
renderer actually committed a real load for the blocked URL and
killing the renderer if that load was disallowed (e.g., for loading
Chrome Web Store in a frame).

mkwst@ is working on a CL
(https://codereview.chromium.org/1617043002/) that will ultimately fix
this by moving XFO enforcement to the browser process and committing
an error page when a load is blocked.  Until then, this is a
short-term fix to change the committed URL for the blocked (blank)
page to urlWithUniqueSecurityOrigin (data:,).

BUG=622385

Review-Url: https://codereview.chromium.org/2096453002
Cr-Commit-Position: refs/heads/master@{#401664}
(cherry picked from commit 30535f7116c9073705a155c7cf4b0146a28f7293)

Review URL: https://codereview.chromium.org/2108873002 .

Cr-Commit-Position: refs/branch-heads/2743@{#509}
Cr-Branched-From: 2b3ae3b8090361f8af5a611712fc1a5ab2de53cb-refs/heads/master@{#394939}