Add GN split_list function.
This function splits a list into N sub-lists. The plan is to use this in a template to automatically split static libraries.

Add a new class ParseNodeValueAdapter which allows getting a list out of a function call without having to copy it. This is used in the new code and is used in a cleanup of rebase_foreach (both of which can take large input lists).

Review-Url: https://codereview.chromium.org/2095043005
Cr-Commit-Position: refs/heads/master@{#402555}