Fix DCHECKs caused by in-process gpu thread.
When run in-process, the GpuSwitchingManager is already notified in GpuDataManagerImplPrivate::HandleGpuSwitch. The call from GpuChildThread causes GpuSwitchingManager to notify its obervers on the GPU thread, while they're expecting it to be UI thread. As a result, a dozen or two of content_browsertests are failing for me in each run (mac debug build).

BUG=

Review-Url: https://codereview.chromium.org/2109643002
Cr-Commit-Position: refs/heads/master@{#402568}