Revert of [Chromecast] Use perceptually linear scaling for stream volumes (patchset #2 id:20001 of https://codereview.chromium.org/2073223003/ )
Reason for revert:
I think we should actually leave the per-stream volume multiplier as linear, and leave it up to whatever sets that multiplier to map correctly from user input to the desired volume curve. Otherwise, we tie the mixer implementation to a specific volume curve, which is looking less and less plausible (particularly since it is used with OEM devices that have an unknown volume curve).

Original issue's description:
> [Chromecast] Use perceptually linear scaling for stream volumes
>
> SetVolumeMultiplier(0.5) should give audio that sounds half as loud as
> SetVolumeMultiplier(1.0).
>
> BUG=internal b/29253848
> TEST=manual
>
> Committed: https://crrev.com/42e51772feb06a971921e4cfcd89ba7661cdf22a
> Cr-Commit-Position: refs/heads/master@{#401149}

TBR=igorc@chromium.org,alokp@chromium.org,halliwell@chromium.org,jyw@chromium.org
# Not skipping CQ checks because original CL landed more than 1 days ago.
BUG=internal b/29253848

Review-Url: https://codereview.chromium.org/2100733002
Cr-Commit-Position: refs/heads/master@{#402591}