Don't lazy allocate the AtomicStringTable.
We can store the AtomicStringTable in a unique_ptr in the WTFThreadData
directly and then use the constructor and destructor to handle adding
and removing the strings. We can then return a reference too since
all WTFThreadData instances have an AtomicStringTable.

Future patches will move most of the AtomicString.cpp logic into the
AtomicStringTable class making the code easier to understand.

Review-Url: https://codereview.chromium.org/2103373002
Cr-Commit-Position: refs/heads/master@{#402684}