PlzNavigate: detect when a ServiceWorker is present
This CL is the first step in PlzNavigate support of ServiceWorker. Before
starting a navigation, it checks whether a ServiceWorker is registered for the
url. If yes, it bails out. Currently this cancel the navigation. The next step
is ot have the navigation sent to the renderer to be handled by the
ServiceWorker.

This CL also removes the code that was introduced as part as the previous
attempt to support ServiceWorker and is no longer needed with the current plan.
BUG=440463
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/2045383002
Cr-Commit-Position: refs/heads/master@{#402791}