Weave Packet Generator This CL adds a packet receiver that respects the uWeave.BLE protocol.
The packet receiver expects this chain of events:
1 control_request/response -> 0+ data packets -> 0 or 1 control_close.

The packet receiver keeps track of the receiving state and will jump to ERROR state when anything unexpected happens. (receiving packets out of order, invalid packet content, etc)

The expected user of packet receiver should feed a packet into receiver, check the state, and take appropriate action. If actions are not taken in the appropriate state, the receiver will crash. Data message should be retrieved promptly when the state reaches DATA_READY or else the data will be lost when the next packet arrives.

Design Doc for ProximityAuth uWeave Migration:
https://docs.google.com/a/google.com/document/d/1HJ1fipxsNQwIxmKs-5_6TjroCAcs6qRhAYvOnIr6mQw/edit?usp=sharing

BUG=617238

Review-Url: https://codereview.chromium.org/2053013002
Cr-Commit-Position: refs/heads/master@{#402884}