Implement the new text-size-adjust CSS property
This patch adds a new CSS property for controlling text autosizing:
text-size-adjust (https://drafts.csswg.org/css-size-adjust).

Intent to implement and ship:
https://groups.google.com/a/chromium.org/d/msg/blink-dev/-vHFK4g93jA/JW9wAJyKAQAJ

At a high level, this patch adds a new CSS property and plumbs it to
storage on the rare inherited style data. The new datatype for this is
"TextSizeAdjust" which just wraps a float with logic for storing both
a floating point adjustment value, and whether 'auto' should be used.
The text autosizer itself has been modified to read this new property.

A few changes were needed to allow the text autosizer to use
multipliers less than 1 (see: TextAutosizer::computeAutosizedFontSize
and callers to that function).

The simple-paragraph.html layouttest has been switched to a unit test
to prove the unit test infrastructure works, and many new tests have
been added for the text-size-adjust feature.

BUG=623158

Review-Url: https://codereview.chromium.org/2100013002
Cr-Commit-Position: refs/heads/master@{#402916}