Fix AW tests for Spitzer
In constast to the MediaPlayerBridge/MediaSourcePlayer
based pipeline Spitzer never creates an underlay SurfaceView
(ContentVideoView) when it works for WebView.

This ContentVideoView was required for some tests. this CL removes
this requirements for all full screen tests and disables
hole punching tests since we consider them obsolete.

This is a prerequisite for enabling Spitzer in Android Webview
(https://codereview.chromium.org/2057743002).

BUG=597495

Review-Url: https://codereview.chromium.org/2029053003
Cr-Commit-Position: refs/heads/master@{#402950}