Expose readonly animation timeline
This patch exposes the Animation.timeline attribute in Animation.idl, as a
readonly attribute with mutability left as a TODO. By exposing this feature, a
few more of the Web Animations web platform tests pass.

Two tests go from failing to flaky:
- imported/wpt/web-animations/interfaces/AnimationEffectTiming/endDelay.html
- imported/wpt/web-animations/interfaces/Animation/playbackRate.html
Where previously these tests failed because timeline was undefined, they now
fail due to values of the timeline sometimes not being as expected, which is a
separate issue.

Something very similar to this patch had previously been submitted in
crrev.com/1942683002 but was reverted due to these flakes.

BUG=600248

Review-Url: https://codereview.chromium.org/2094213002
Cr-Commit-Position: refs/heads/master@{#403051}