content: convert QuotaPolicyCookieStoreTest to use CanonicalCookie's Create() method.
This patch converts the usage of CanonicalCookie's constructor in
QuotaPolicyCookieStoreTest tests to use CanonicalCookie::Create() method.

We are doing this because the Create() method validates the parameters
and we thus we can test the return value to see if it has passed or not.
Whereas we can not do this when using the constructor.

BUG=57061
TEST=content_unittests
R=mmenke@chromium.org,avi@chromium.org

Review-Url: https://codereview.chromium.org/2111533002
Cr-Commit-Position: refs/heads/master@{#403109}