Add ink drop highlight for hover and focus on location icons (e.g. star, zoom, translate, passwords) and content settings icon.
This doesn't do anything for other icon label bubble views (like the
location icon). Some of the code in ContentSettingImageView should
probably be moved into IconLabelBubbleView for that. Also, there's no
hover or focus effect when the label is showing (e.g. when Chrome first
blocks a popup and there's text next to the icon). Since it remains
unclear to me how the hover and press effects should look in this
scenario, limit the ink drop to when only the icon is visible.

BUG=591140

Review-Url: https://codereview.chromium.org/2098643002
Cr-Commit-Position: refs/heads/master@{#403170}