arc: Notify the user on fatal opt-in flow errors
If ARC has a fatal error during first initialization, the user will
never be notified. It might also cause a case where ARC would be stopped
and started infinitely without giving the user a chance to stop it. This
change adds some logic to stop automatically retrying to start ARC and
let the user decide to do so instead.

BUG=b/29833836
TEST=Kill ARC mid-initialization, "Service unavailable" is shown

Review-Url: https://codereview.chromium.org/2114463002
Cr-Commit-Position: refs/heads/master@{#403231}