[Remoting Android] Make JniClient own JniDisplayHandler
Refactorization CL preparing for OpenGL rendering on Android. Currently
JniDisplayHandler is owned and managed by the Java display object, which
requires passing pointer of the display handler back and forth and makes the
lifetime management of the display handler not very straightforward. This CL
makes JniClient own JniDisplayHandler and makes JniDisplayHandler control
the lifetime of the Java display object instead of the other way around.

BUG=385924

Review-Url: https://codereview.chromium.org/2100943004
Cr-Commit-Position: refs/heads/master@{#403272}