Replace "Remove from Chrome..." in the action context menu with "Remove icon"
The toolbar action for Media Router has a context menu option to "Remove from
Chrome...", which doesn't show a confirmation dialog like other extension icons
with the same option do (and ellipsis usually suggests a confirmation). Also,
since the removal merely removes the icon and none of the MR functionalities,
the use case is different from the removal of other extensions on the toolbar.
This CL replaces the text with "Remove icon".

BUG=624104

Review-Url: https://codereview.chromium.org/2112623002
Cr-Commit-Position: refs/heads/master@{#403314}