Consider non-keyframes for RTCVideoEncoder error counter
This CL addresses the bug below. Earlier on, we experimented and
decided that we would fallback to SW decoded after 50 consecutive
HW decode errors. However, after an error in keyframe, non-keyframes
would be dropped and not added to the error counter. As a result,
we would wait for 50 keyframe errors before SW fallback, which
takes a long time. This CL fixes the issues resulting from this.

BUG=616968
TEST=Tried the problematic Cisco stream referred in bug.
First 3-5 seconds, we don't get a video(due to HW errors) but works
afterwards(falling back to SW).

Review-Url: https://codereview.chromium.org/2093823002
Cr-Commit-Position: refs/heads/master@{#403353}