Support dragging texts into Android WebView.
WebView gets callback 'onDragEvent' when something is dragged into the WebView.
This CL extracts the texts (if any) and location from the DragEvent and creates
DropData which is Chromium's internal representation of dragged data.

The event plumbing path is WebView#onDragEvent --> AwContents#onDragEvent -->
ContentViewCore --> WebContentsViewAndroid.

This CL also uses a new blink::WebView API that performs dropping with DropData,
because on Android, the DropData is only available when the drop is performed,
not when the drag is started.

BUG=584789

Review-Url: https://codereview.chromium.org/1728193002
Cr-Commit-Position: refs/heads/master@{#403417}