Fix crash when variation parameters are set in chrome://flags.
In chrome://flags, one can now choose among predefined variation
parameters for a feature. Such parameters are registered on startup
with a predefined trial and group name. This causes a crash when the
same trial is specified in Finch config.

This CL skips dealing with a Finch trial where a group is already
selected and this group is not mentioned in the Finch config.

BUG=624807

Review-Url: https://codereview.chromium.org/2119713002
Cr-Commit-Position: refs/heads/master@{#403431}