Add IPC serialization support for base::Optional
As part of bug 623556, we need to update one of our IPC structs to use
Optional.

We unfortunately can't migrate to mojo at this time because our IPCs
depend on delivery order relative to other IPCs (in particular,
FrameHostMsg_DidCommitProvisionalLoad). Additional details on ordering
constraints are covered in
https://docs.google.com/document/d/1HJsJ5W2H_3qRdqPAUgAEo10AF8gXPTXZLUET4X4_sII/edit.
As a result, I'd like to add IPC serialization support for Optional
so we can unblock progress on bug 623556.

BUG=623556

Review-Url: https://codereview.chromium.org/2098823003
Cr-Commit-Position: refs/heads/master@{#403442}