Fix flaky Kiosk browser_tests failing on wrong widget opacity.
Ash was trying to show the shelf with its widget opacity set to 0.0 causing
a DCHECK crash in the tests. This makes sure the opacity of the system tray's
widget is set properly when it's shown or hidden.

Also re-enables a test that was disabled previously since it's passing now due
to this CL.

BUG=624584
TEST=browser_tests --gtest_filter=KioskEnterpriseTest.PrivateStore
TEST=browser_tests --gtest_filter=KioskUpdateTest.IncompliantPlatformFirstInstall

Review-Url: https://codereview.chromium.org/2119913002
Cr-Commit-Position: refs/heads/master@{#403489}