ManageSpace activity - Do not load browser on insta crash.
Use a preference to track whether the managed space activity
successfully completes before attempting to load the browser
process on the next startup.

This should ensure that native crashes in the renderer process
do not prevent users from clearing their data.

It uses the current product version to allow the browser process
to be attempted after upgrades regardless of previous failures.

BUG=571461

Review-Url: https://codereview.chromium.org/2111813003
Cr-Commit-Position: refs/heads/master@{#403553}