[Extensions UI] Move search settings bubble to be cross-platform
Instead of triggering the search settings changed notification bubble
from the platform-specific OmniboxViewViews, add an observer on the
cross-platform OmniboxView so that it can be handled on both windows
and mac. Also rework settings_api_bubble_helpers to be more OO.
Port existing views browsertests for search settings and controlled
home notifications to mac.

BUG=497340

Review-Url: https://codereview.chromium.org/2093843002
Cr-Commit-Position: refs/heads/master@{#403592}