cc: Don't skip clip rect calculation for copy request layers
When a layer with copy request is SetMasksToBounds, it is considered
clipped, but the clip rect calculation is skipped, thus resulting in
empty clip rect. This CL fixes this case.

In production this case was never hit because for root layers we don't
care about clip rect, and for non root copy request we don't
SetMasksToBounds.

BUG=
R=ajuma
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review-Url: https://codereview.chromium.org/2117173002
Cr-Commit-Position: refs/heads/master@{#403723}