Don't require AMP URLs from new backend.
Without this change, we can fail to process snippets from the Chrome
Content Suggestions server. Unfortunately, that means that the whole
response is discarded, rather than just the particular problem snippet
(maybe we should reconsider this?).

If we want to test with the new server in M53, we have a two options:

  * Update the server to produce only responses that are accepted by the
    current NTPSnippet implementation. This means we'd have to either:
    * provide a URL like "about:blank", which might have bad
      side-effects, or
    * drop them entirely, which wouldn't be great for our coverage.

  * Merge this CL into M53.

BUG=621090

Review-Url: https://codereview.chromium.org/2113313002
Cr-Commit-Position: refs/heads/master@{#403766}