Reland: Make 2D canvas disable gpu acceleration when getImageData is called
In order to optimize performance, GPU-accelerated canvases should
permanently switch out of accalerated mode as soon as
getImageData is called.  The cost of getImageData on the GPU is
so prohibitively high that it almost always outways the benefit
of acceleration.

This is the first time we implement a SW/GPU switch that can happen
after the first frame was presented, which may result in a small one
time glitch due to rendering engine discrepancies. Let's see if we get
any complaints about this.  If there are complaint, we may inhibit
the switch based on what kind of content was rendered to the canvas.

BUG=606688

Review-Url: https://codereview.chromium.org/2123623002
Cr-Commit-Position: refs/heads/master@{#403806}