Use RelocatablePosition in CompositeEditCommand::moveParagraphs()
Currently, CompositeEditCommand::moveParagraphs() uses a pair of
VisiblePositions to keep track of original position of the moved
paragraph after it's moved away. However, the VisiblePositions may
be invalidated if their anchor nodes are moved out of document,
causing crashes when the VisiblePositions are used later. It causes
the failure of layout test:

fast/text/first-letter-bad-line-boxes-crash.html

This patch uses RelocatablePosition for the tracking purpose as
a bug fix.

BUG=581038
TEST=fast/text/first-letter-bad-line-boxes-crash.html

Review-Url: https://codereview.chromium.org/2127503002
Cr-Commit-Position: refs/heads/master@{#403879}