Fix blimp native Java tests on platforms with assert.
The BlimpNativeInstrumentationTestCase wrapper for blimp tests that need
to load native code to run was not using the UI thread for loading the
library. There was an assert checking for this, but this doesn't execute
on newer platforms, so was not caught during development or in the CQ.

This CL fixes the issue by ensuring that the library is loaded on the
correct thread.

BUG=622236

Review-Url: https://codereview.chromium.org/2124143002
Cr-Commit-Position: refs/heads/master@{#403965}