mus: Add an API for client initiated window moves.
This adds a set of APIs so the client can perform a blocking window
move. The client asks the window server to perform a window move, which
forwards this request to the window manager. We want to do this instead
of having the client perform all moves itself because the window manager
performs special actions, like docking a window when dragged to the
side.

BUG=622900

Review-Url: https://codereview.chromium.org/2060513002
Cr-Commit-Position: refs/heads/master@{#404021}