Reset translation of toolbar_buttons for Url Defocus Animation.
The mToolbarButtonsContainer has its translationY modified during scroll so
when the user scrolls on the NTP, it appears to scroll too. However during
the URL focus and defocus animations it should not be touched. Unfortunately
updateNtpTransitionAnimation() is called a few times after the URL focus
animation has been completed while mUrlFocusChangeInProgress is set to false,
causing translationY to be non-zero at the end.  We reset the translationY
here so the mToolbarButtonsContainer is on screen for the defocusing animation.

BUG=623885

Review-Url: https://codereview.chromium.org/2121283002
Cr-Commit-Position: refs/heads/master@{#404113}