Make PlatformScreenImplOzone use DisplayConfigurator.
Currently the PlatformScreen implementation for Ozone uses
NativeDisplayDelegate directly. DisplayConfigurator is already designed
to use it and is accessible from services/ui. Change the implementation
to use DisplayConfigurator instead.

This provides PlatformScreen with much of the necessary display
configuration code it will need for multiple monitor support. For now
PlatformScreenImplOzone just forces the native resolution on 1 display.

BUG=625876

Review-Url: https://codereview.chromium.org/2121053002
Cr-Commit-Position: refs/heads/master@{#404151}