Enable the WebRTC event log on PeerConnection instead of PeerConnectionFactory.
This allows simultaneous recording of event logs of multiple PeerConnections, each in a seperate file. The number of files is limited, as well as the size of each file. A callback handler is added to RenderProcessHost to allow code in chrome/ to register callbacks to implement the WebRTC event log functionality.

BUG=chromium:600661,chromium:613499,webrtc:4741

Review-Url: https://codereview.chromium.org/1855193002
Cr-Commit-Position: refs/heads/master@{#404183}