Move content's shell connections to the IO thread
Moves all shell connections in the content layer
(connections owned by instances of MojoShellConnection and
MojoChildConnection) to the IO thread in all process types.
Note that instances of MojoShellConnection and
MojoChildConnection may still be created and used on
any thread.

Additionally exposes a way for InterfaceRegistry and
InterfaceProvider to forward to/from other instances
so we can still have a corresponding instance of each
on whatever thread owns the Mojo[Shell/Child]Connection
object.

This simplifies interface registration and acquisition
on process hosts and ChildThreadImpls, and makes it possible
for a future CL to introduce IO-thread service registration
to avoid thread hops when binding incoming interface requests.

BUG=623398,612500

Review-Url: https://codereview.chromium.org/2111353002
Cr-Commit-Position: refs/heads/master@{#404228}