Condition LoFi field trial on LoFi flags not set
The session is considered to be in LoFi field trial
(Control or Enabled) only if Lo-Fi flag is in default
state.

This removes the ambiguity of what logic to apply when the
session is in field trial, and LoFi is enabled from the
flags, and makes it easier to write related code. e.g., we
can condition reporting Lo-Fi UMA only when the session is
in the field trial without checking for the flags.

BUG=626138

Review-Url: https://codereview.chromium.org/2128293002
Cr-Commit-Position: refs/heads/master@{#404242}