MD Settings: Reset page, fix dialog test that was erroneously passing.
The test that checks if a dialog can be closed in all possible ways ('x' button,
cancel button, escape key), was running the three sub-tests on the same dialog
instance in parallel. This caused all three sub-tests to pass, even though when
ran individually the 'escape' key test was failing.

Resolved by changing tests to not run in parallel, and fixing test code to
properly test closing using the 'Esc' key.

This bug was revealed while investigating migration to native <dialog>.

BUG=625332

Review-Url: https://codereview.chromium.org/2130023002
Cr-Commit-Position: refs/heads/master@{#404276}