[Chromecast] Update media time when playback is started
When CMA media pipeline is started we post a task that will continuouly
update media time. But since that task is posted, it won't run right
away. In the meantime somebody might call GetMediaTime and will get
an invalid value (kNoTimestamp). To avoid this we can just set the
last_media_time_ right away for the first time when starting playback.
This will ensure that we have valid media time even before the first
UpdateMediaTime task invocation.

BUG=internal b/30035825

Review-Url: https://codereview.chromium.org/2131413002
Cr-Commit-Position: refs/heads/master@{#404513}