Remove dtlsidentitystore.h from libjingle GYP/GN files.
This file contains an interface that is about to be removed
(https://codereview.webrtc.org/2034013003) since it is no longer used
by chromium (https://codereview.chromium.org/2033353002).

The build files have to be updated before the file is removed from the
webrtc repo or else the webrtc roll will cause chromium to break.

BUG=webrtc:5708, webrtc:5707
NOTRY=true

Review-Url: https://codereview.chromium.org/2120493002
Cr-Commit-Position: refs/heads/master@{#404614}