Disable compositing asserts when gesture scrolling on main thread
These asserts get tripped when ScrollAnimator tries to determine if the
scroller needs to be scrolled on the main thread or on the impl thread.

Also moved the early-out if there's no FrameView on GestureScrollBegin
to all GestureScroll events since we shouldn't handle any of those if
there's no view.

BUG=625676

Review-Url: https://codereview.chromium.org/2120373002
Cr-Commit-Position: refs/heads/master@{#404922}