Remove WebContents::HasManifest.
WebContents::GetManifest now returns a manifest URL and a manifest
object. The non-emptiness of the URL correlates to whether HasManifest
returned true or false. This CL removes the HasManifest method from
WebContents and refactors its sole client, app banners, to check the
manifest URL returned in GetManifest instead. This eliminates a roundtrip
IPC from the app banner flow, which runs on every page load on mobile.

Minor changes to when the manifest URL is set in the manifest parsing
process are also made to ensure the correctness of this approach.

BUG=626969

Review-Url: https://codereview.chromium.org/2140463002
Cr-Commit-Position: refs/heads/master@{#404973}