MD WebUI: Reimplement cr-shared-menu using iron-dropdown.
This change brings the implementation much closer to that of
paper-menu-button, and fixes a number of longstanding bugs. In
particular,

* Focus is correctly restored to the element which opened the menu after
  the menu is closed.
* It is only necessary to press escape once to close the menu.
* It is possible to close the menu by clicking anywhere else on the
  page.
* Tab focusing is trapped inside the menu, rather than allowing tab to
  go to a different part of the page.
* Bold text is no longer applied the menu item which was last clicked.

The main trade-off of this approach is that we now have to use neon-animation
to animate the menu out, and due to the way that the menu is positioned
it will animate in from the top-left corner (the same as
paper-menu-button).

BUG=589362,603454,616074,619839,620038
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2104013004
Cr-Commit-Position: refs/heads/master@{#405048}