Add ContentSuggestionsService to SnippetsInternals
The ContentSuggestionsService retrieves suggestions from the
NTPSnippetsService as a provider. The latter currently delivers its
snippets to the chrome://snippets-internals/ page for debugging.
Before the NTP UI will be changed to retrieve suggestions from the
ContentSuggestionsService instead of directly from the
NTPSnippetsService, this adds the output of the
ContentSuggestionsService to the snippets-internals page to allow
debugging both and compare the data converted data to the original.
This also adds controls for clearing cached/discarded suggestions.

BUG=619560
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2145563002
Cr-Commit-Position: refs/heads/master@{#405149}