Clean-up the code for default variations in chrome://flags.
The CL 2129543002 introduced an option to override variation parameters
using a feature in chrome://flags. The list of options for the given
feature always includes "Default", "Enabled", and "Disabled". Even
though the semantics of "Enabled" was meant as 'enabled with the default
parameter values', the API forced to specify the variation for "Enabled"
as well.

This is unnecessarily cumbersome and confusing. This CL makes the
"Enabled" option completely automatic - no need to specify its variation
any more.

BUG=609059

Review-Url: https://codereview.chromium.org/2146903002
Cr-Commit-Position: refs/heads/master@{#405190}