Fix crash when destroying a RenderWidgetHost that holds the pointer lock.
This CL fixes two separate issues. One is that when a
CrossProcessFrameConnector is being destroyed, if the
RenderWidgetHostViewChildFrame holds the mouse lock it must
unlock the mouse.

The second issue is that when flash holds a fullscreen
pointer lock and releases it, the WebContents is notified
of the lost mouse lock, even though it was never notified
that the mouse lock was acquired (because flash has extra
privileges to acquire the lock).

BUG=619571
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/2112923002
Cr-Commit-Position: refs/heads/master@{#405195}