Stop tracking user input events without a relevant load.
These events are 100x more common than other errors. User input events
can arrive when the user is e.g. looking at an image, which isn't a
tracked page but is something a user can interact with, so it's valid
to receive input events without a committed load. Thus, it doesn't make
sense to log this error.

Review-Url: https://codereview.chromium.org/2142743002
Cr-Commit-Position: refs/heads/master@{#405286}