Clean up some OOPIF-related naming and comments in Fullscreen.
m_forCrossProcessAncestor was intended to mean that fullscreen was
entered for an iframe container which is an ancestor of the actual
fullscreen element, which also mirrors the
:-webkit-full-screen-ancestor style that the container will gain.  But
on the flipside, this naming can also be confusing: if an ancestor of
a cross-process iframe goes fullscreen, that iframe doesn't need to
know about this at all, and instead it only needs to know when an
element in a descendant frame goes fullscreen.  It might make more
sense to name this flag to reflect where the actual fullscreen element
is, rather than how m_fullscreenElement relates to it, so this CL
renames the flag to m_forCrossProcessDescendant and updates comments.

BUG=550497

Review-Url: https://codereview.chromium.org/2147803002
Cr-Commit-Position: refs/heads/master@{#405287}