Re-reland: Add MemoryCoordinator
This is second attempt. Hope GYP Mac is happy with this :(

Original description:

This CL adds MemoryCoordinator mojo interface and
its implementation. When --enable-features=MemoryCoordinator
is specified the browser and renderers create their
MemoryCoordinators. MemoryCoordinatorImpl (which lives in
the browser process) creates a mojo message pipe when a
renderer is launched.

This CL doesn't add actual handling of memory events, which
will be implemented in follow-up CLs.

BUG=617492
TBR=chrisha@chromium.org,haraken@chromium.org,dcheng@chromium.org,jam@chromium.org

Review-Url: https://codereview.chromium.org/2149063002
Cr-Commit-Position: refs/heads/master@{#405460}