Add callback list to PrefModelAssociator after sync data is loaded
This is a follow up CL of https://codereview.chromium.org/1895993003. The
migrated settings will be synced back because sync process happens after the
construction of HostContentSettingsMap.

In this CL, a callback list is added to PrefModelAssociator, which will gets
registered with the migration code after construction of HostContentSettingsMap,
and gets executed at the end of PrefModelAssociator::MergeDataAndStartSyncing.

The follow up CL of this is https://codereview.chromium.org/2075103002.

BUG=604612

Review-Url: https://codereview.chromium.org/2078893002
Cr-Commit-Position: refs/heads/master@{#405669}