MD WebUI: Center search field correctly inside cr-toolbar
This changes the behavior of the search field so that it no longer
expands on open in wide windows. This makes it possible to center the
search field inside the toolbar. Also tweaks the open/close animation
for the toolbar in narrow mode to be smoother.

BUG=610609
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:closure_compilation;master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2140943002
Cr-Commit-Position: refs/heads/master@{#405681}