Fix apps page icon in tabstrip.
Note that the apps button on the bookmark bar works correctly, but if you
also bookmark the apps page, the bookmark will be theme-colorized
although it probably shouldn't. However, I consider this much more of a
corner case than the described bug (which occurs any time you're using the
app launcher with a themed browser). So for now, just fix the tabstrip
problem.

BUG=628599

Review-Url: https://codereview.chromium.org/2150423002
Cr-Commit-Position: refs/heads/master@{#405827}