Show download error message if sdcard is not available
Currently when downloading a file without sdcard, Chrome shows the overwrite
infobar.
And if user acks that infobar, the file will be downloaded to Chrome's own
dir.
And the downloaded file cannot be opened by other apps, making it
useless.
This CL fixes that behavior by showing a "missing sdcard" message
when sdcard is not available.
Chrome will no longer try to generate a path under its own dir if sdcard
is not available.
If snackbar is available, the error message will be displayed on
snackbar.
Otherwise, a toast will be shown.

BUG=624912

Review-Url: https://codereview.chromium.org/2117343007
Cr-Commit-Position: refs/heads/master@{#405833}