DevTools: [SSP] render selectors effectively
Currently, the selectors in styles sidebar pane are rendered each in a
separate span, which makes them ctrl-clickable.

This is needed to help in SourceMap debugging, as different selectors
might be coming from different places in sources. (see crbug.com/266416)

However, as it generates a lot of DOM nodes, this patch stitches unmatched
selectors together in a single element. The selectors, however,
are rendered separately as the user hits Ctrl and hovers over the section.

BUG=628409
R=pfeldman, dgozman

Review-Url: https://codereview.chromium.org/2146233003
Cr-Commit-Position: refs/heads/master@{#405934}