Split redirect delay metrics into separate histograms.
In https://codereview.chromium.org/2103153004/, a histogram was added
to track the time between a page painting, and initiating a client-side
redirect.

Additionally, a zero value was recorded in the following two cases:
* the page initiated a client-side redirect, then painted
* the page initiated a client-side redirect, but never painted

It turns out to be important to distinguish these cases. Distinguishing
these cases will let us understand how frequently a page may initiate a
redirect, paint, then have the redirect commit. In this case, we may want
to suppress the paint notification since it ends up being more of an
interstitial paint while navigating to the actual destination page, rather
than what we'd consider an actual first paint. In order to decide whether
we should add logic to suppress, this change separates the histogram into
two so we can understand how common this case is.

BUG=613634

Review-Url: https://codereview.chromium.org/2151363003
Cr-Commit-Position: refs/heads/master@{#406192}