Disallow user overrides to enterprise policy for plugins.
This CL changes the code in two places:

Firstly, in the actual plugin loading logic, in
PluginInfoMessageFilter::Context::GetPluginContentSetting
there was previously an exception for CONTENT_SETTING_ASK
which meant that it could override enterprise policy.

Secondly, in the UI, this disables the tick box to "Always
allowed to run" on chrome://plugins when enterprise policy
is CONTENT_SETTING_ASK (3).

BUG=616218
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2150323003
Cr-Commit-Position: refs/heads/master@{#406197}