Add a mojom/StructTrait for display::Display.
There is currently a mojom that mostly corresponds to display::Display
and a corresponding TypeConverter. However, the existing mojom has some
extra information that is mus specific. In the future ui.mojom.Display
will be modified to contain a display.mojom.Display to avoid
duplication.

This CL only adds the new mojom, StructTrait and unit tests to ensure
StructTrait works correctly.  The new mojom and StructTrait will be
needed for the DisplayController Mojo interface to allow display
configuration to happen. Unit tests will be added to trybots in a
follow-up CL.

BUG=612242

Review-Url: https://codereview.chromium.org/2123613002
Cr-Commit-Position: refs/heads/master@{#406319}