Always show Chrome app on IntentPickerBubbleView
Swapping results on ArcNavigationThrottle so we always show the Chrome
Browser app to the user without the need of scrolling down on the UI.
Exposing |kMaxAppResults| so IntentPickerBubbleView can use it.

BUG=622384
TEST=Having installed a handful of third party browsers plus the
Pinterest app on ARC go to www.future-cities-lab.net and click the
Pinterest button, you must see the Chrome Browser as an option on the
top 3 regardless of how many Android apps you have installed.

Review-Url: https://codereview.chromium.org/2157663003
Cr-Commit-Position: refs/heads/master@{#406438}