Don't create zero-velocity GestureFlingStarts.
Currently trackpad initiated GestureFlingStarts on CrOS can have zero-
velocity, but this causes crashes in the renderer. This CL converts
these events to be GestureFlingCancels (just in case they are issued
in the middle of ongoing fling).

As it is hard to find the sources of such events after the fact, a
DCHECK has been added to catch new pathways if they issue such events.

BUG=627227

Review-Url: https://codereview.chromium.org/2165043002
Cr-Commit-Position: refs/heads/master@{#406678}