Mac (Cocoa) Omnibox: Force text field to LTR context if it is a URL.
This means that URLs will be displayed in a left-to-right paragraph
context. Right-to-left runs are still rendered RTL, but will not flip
the whole URL around. For example (if "ABC" is Hebrew), this will render
"ABC.com" as "CBA.com", rather than "com.CBA".

Affects main text field and suggestions, but not non-URL search text.
Complies with RFC 3987 Section 4.1 and brings the Mac Omnibox in line
with the existing behaviour on Views and Android.

BUG=624213

Review-Url: https://codereview.chromium.org/2126023003
Cr-Commit-Position: refs/heads/master@{#406791}