[iOS] Refactor iOS bots definitions.
Refactor ios-{device,simulator}{,-gn}.json files so that the difference
between gyp and gn bot is limited to the "comment" and "mb_type" fields.

Sort "gn_args" values so in the same order as the corresponding value
in "GYP_DEFINES" (this make it easier to check that both gyp and gn are
using the same settings).

Convert ios-simulator-gn.json to build a fat binary and run the tests
on both a 32-bit and a 64-bit device as ios-simulator.json is doing.

BUG=621708

Review-Url: https://codereview.chromium.org/2162133003
Cr-Commit-Position: refs/heads/master@{#406869}