[Media Router] Adds return value to mojo MediaRouteProvider::TerminateRoute.
This API change allows the MRPM to pass a Promise to the MR that represents the
outcome of a call to terminateRoute().  The result is logged in a new histogram, MediaRouter.Provider.TerminateRoute.Result.

The change is designed to be backwards compatible with MRPM versions that do not return a Promise.

The MRPM side change will be done next.

BUG=627967

Review-Url: https://codereview.chromium.org/2145983003
Cr-Commit-Position: refs/heads/master@{#406897}