Replace gfx::Rect with gfx::RectF in ui::AXNodeData
This is a mechanical change to switch to using floating-point rects internally
for accessible bounding boxes.  In cases where we need to convert back to
a gfx::Rect, we use gfx::ToEnclosingRect.

This change is necessary so that in the future we can use relative
bounding boxes and transformation matrices in the accessibility tree.
In isolation this change should have no effect except that objects might
be off by one pixel due to differences in when rounding happens.

BUG=618120

Review-Url: https://codereview.chromium.org/2154213007
Cr-Commit-Position: refs/heads/master@{#406963}