DevTools: cleanup treeoutline expand logic
TreeElement.expand() should not need to check that the element has children and
TreeOutline.insertChild() should not automatically expand a child when it is
inserted. ObjectPropertiesSection.js depended on this logic to automatically
expand the Entries property. This CL updates it so that the Entries property is
added before expanded.

BUG=

Review-Url: https://codereview.chromium.org/2147343002
Cr-Commit-Position: refs/heads/master@{#407061}