Open the notification URL (not the web app start_url) when deep-linking into standalone mode.
When deep-linking into a standalone web app from a notification, the
app's start_url is always opened and the URL that the notification wants
to open is ignored. This CL fixes the bug by injecting the notification
URL in place of the web app URL when deep-linking. We also force a
reload of the web app if it is already running to ensure that the desired
url is loaded.

Replacing the URL in this manner is correct as the notification URL is
guaranteed to live within the web app's scope for notification
deep-linking to trigger.

BUG=627791

Review-Url: https://codereview.chromium.org/2145833003
Cr-Commit-Position: refs/heads/master@{#405638}
(cherry picked from commit e10048ec434da0f56f6b2c5579bdbee2fa06654d)

Review URL: https://codereview.chromium.org/2173643002 .

Cr-Commit-Position: refs/branch-heads/2785@{#281}
Cr-Branched-From: 68623971be0cfc492a2cb0427d7f478e7b214c24-refs/heads/master@{#403382}