Change NTPSnippetsBridge to read from ContentSuggestionsService
Change the NTPSnippetsBridge to receive ContentSuggestions from the new
ContentSuggestionsService instead of receiving NTPSnippets directly
from the NTPSnippetsService. Change the bridge/UI to receive and handle
the ContentSuggestionsCategoryStatus instead of the state and
DisabledReason from the NTPSnippetsStatusService.

Modify the implementations of NTPSnippetsService
::DiscardSuggestion and ::FetchSuggestionImage to accept
combined unique_ids instead of within_category_ids, which they
wrongly accepted before. The new implementations conform to
the ContentSuggestionsProvider interface and these two methods
should not be called directly on the NTPSnippetsService with
an ID of a NTPSnippet anymore (as it would now fail).

BUG=619560

Review-Url: https://codereview.chromium.org/2158883002
Cr-Commit-Position: refs/heads/master@{#407441}