[Cronet] Mark request as complete when OutputStream fails
When consumer specifies a chunked upload or a fixed stream upload, Cronet's
HttpURLConnection starts the request right away. If an error happens, we will
surface that error as an IOException to consumer. However, if the consumer
tries to call any other API (like getResponseCode(), getResponseMessage())
afterwards, we will spin up the message loop and wait forever.

This CL marks the request as complete if an error occurred in upload, so
getResponseCode() etc. do not spin up the message loop to wait for response.

R=pauljensen@chromium.org
BUG=630664

Review-Url: https://codereview.chromium.org/2173923002
Cr-Commit-Position: refs/heads/master@{#407468}