Fix networking_private IDL
Minor fixes to the networkingPrivate IDL.
This produces a non-fatal but confusing error in the
Settings UI.

ManagedDOMString[] is invalid, ManagedDOMStringList
should be used instead. (The managed property should
be an objetc with each entry, e.g. 'Active', 'DevicePolicy',
etc, as a separate array of strings, instead of an array of
objects).

BUG=none

Review-Url: https://codereview.chromium.org/2182633002
Cr-Commit-Position: refs/heads/master@{#407583}