Add Apple IST CA 8 to the Symantec exclusion, remove CA3 - G1 and NTT
Symantec originally indicated NTT DoCoMo was indepedently operated,
but subsequent investigation revealed it was part of Symantec's MPKI.
Apple's G3 is only audited to WebTrust for CAs, not SSL BRs, so removing
it from the exclusion set.

This adds Apple's CA 8 to the set, after communicating with Apple's PKI
team to receive confirmation that it was created pursuant to the
audited CP/CPS and is part of the same infrastructure.

BUG=620178

Review-Url: https://codereview.chromium.org/2177943004
Cr-Commit-Position: refs/heads/master@{#407618}