Using the uWeave packet generator and receiver to re-implement the BluetoothLowEnergyConnection class
Major changes:
SendMesssageImpl which instead of using the old protocol, it will just generate the message with the WeavePacketGenerator.

GattCharacteristicValueChanged instead of parsing messages received with the old protocol, will now use the WeavePacketReceiver and just switch on the state of the receiver and take appropriate action.

Design Doc for ProximityAuth uWeave Migration:
https://docs.google.com/a/google.com/document/d/1HJ1fipxsNQwIxmKs-5_6TjroCAcs6qRhAYvOnIr6mQw/edit?usp=sharing

BUG=617238

Review-Url: https://codereview.chromium.org/2075313002
Cr-Commit-Position: refs/heads/master@{#407655}