Fix a threading issue and modernize the use of scoped_refptr in the on-demand update checker.
The vector of delegates_ was previously being walked on the FILE
thread. This could lead to a crash if a new delegate were to be added on
the UI thread. Now, the driver bounces back to the UI thread to send
progress updates to the delegates.

Also changes the use of scoped_refptr to be up-to-date with current
style.

BUG=none
R=pmonette@chromium.org

Review-Url: https://codereview.chromium.org/2176123002
Cr-Commit-Position: refs/heads/master@{#407726}