Apply autofill profile use-date fix after sync starts.
The use-date fix replaces invalid autofill profile last-use dates (prior
to M53, the use-date at creation was set to 0). It is currently run when
the autofill profile data is first loaded upon chrome start. This
happens before sync, if enabled, has started. Unfortunately, when sync
starts, it overwrites the fixed last-use dates for any profiles already
represented in the sync data stream.

This CL moves the invocation of...

  autofill::PersonalDataManager::ApplyProfileUseDatesFix()

... to be invoked either upon determining that sync is not enabled
(and will therefore not clobber the fixes) or upon notification
that sync has started (initial sync is complete, and sync is
ready to accept and propagate our fixes).

BUG=631141
R=mathp@chromium.org, sebsg@chromium.org, zea@chromium.org

Review-Url: https://codereview.chromium.org/2177253003
Cr-Commit-Position: refs/heads/master@{#407859}