Switching out of keyword mode shouldn't count as a minimal change.
Previously, if you used space to get into keyword mode and later backspace at
the beginning of the keyword to get out of keyword mode, the backspace would
trigger an autocompletion run with |minimal_changes| == true.  This isn't really
correct; switching out of keyword mode can affect the results and scoring
significantly.  In certain cases, the old route triggered a DCHECK, because
the search provider didn't wipe the old keyword matches across
|minimal_changes|, but the new input had no corresponding keyword input text.

I looked briefly at leaving this as a "minimal change" and instead detecting and
handling it inside the relevant provider(s), but decided that was error-prone
and not really as semantically accurate.  Plus the potential additional flicker
here is fairly minor, as people don't usually exit keyword mode via backspace in
this way.

BUG=607379
TEST=In a debug build, type "google.comgoogle.com" (no quotes) into the omnibox, place the cursor between the two "google.com"s, hit space, then hit backspace.  Chrome should not DCHECK.

Review-Url: https://codereview.chromium.org/2179063003
Cr-Commit-Position: refs/heads/master@{#407888}