[CacheStorage] Deleting a cache could delete the wrong directory
The CacheStorageCache::CacheLoader maps cache names to paths. The problem is
that as of recent changes, it's possible to have two caches with the same name
exist simultaneously (the doomed one and a new one). If you then delete the
doomed one from disk, the CacheLoader will delete the new one's directory
instead since it maps name to path.

This CL fixes the above problem by keeping track of doomed caches and their
paths separately from live caches.

BUG=630036

Review-Url: https://codereview.chromium.org/2186433004
Cr-Commit-Position: refs/heads/master@{#407920}