Turn on enable begin frame scheduling by default
This turns on --enable-begin-frame-scheduling for all platforms but mus
and Blimp.  This change means browser->renderer begin frame ticks
instead of sending vsync information and having a synthetic source on
the renderer side.  This feature was already on for Android so should
only be a real change for desktop / ChromeOS platforms.

Lots of cleanup can follow from this like removing all commit vsync /
authoritative vsync / CompositorVSyncManager things, but this is a
smaller patch to suss out any performance regressions.

This previously landed and regressed a number of metrics.  Some of these
have been investigated and fixed and the rest are WontFix.

CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review-Url: https://codereview.chromium.org/1939253002
Cr-Commit-Position: refs/heads/master@{#408172}