Do not persist HttpServerProperties to disk that often.
Add some heuristics to SetAlternativeServices that prevents a
ScheduleUpdatePrefsOnNetworkThread() call after each incoming Alt-Svc header,
based on the time of the last update, and the expiration times of the current
and new in-memory entries.  (Note that the current in-memory entry might be
different from the entry last persisted to disk.)

Repeated requests to the same server typically result in identical Alt-Svc
announcements, which become different when the max-age field (in seconds from
now) gets converted into a base::Time expiration time.  Alternative services do
not need to be persisted to disk with every incoming Alt-Svc header if other
fields (scheme, host, port) are identical.

BUG=554643

Review-Url: https://codereview.chromium.org/2171743002
Cr-Commit-Position: refs/heads/master@{#408273}