Implement multi-phase build support in MB (merge to M53)
The PGO builders need to run multiple compiles in a single build,
each with different arguments. This was not previously possible
with MB, but this CL adds support for specifying a list of configs
per builder, and adds a --phase command line arg to indicate which
phase of the build should be generated.

It is an error to specify --phase on a builder that isn't
expected to have multiple phases. Also phases are 1-based (phase 1,
phase 2, etc.).

Also, this CL cleans up some TODOs that I no longer plan to fix.

TBR=sebmarchand@chromium.org
BUG=595947

Review URL: https://codereview.chromium.org/2188483005 .

Cr-Commit-Position: refs/branch-heads/2785@{#382}
Cr-Branched-From: 68623971be0cfc492a2cb0427d7f478e7b214c24-refs/heads/master@{#403382}