Removing DownloadInfoTest
It doesn't make sense to have DownloadInfoTest with upcoming ProGuard
changes. It doesn't check any meaningful functionality (just that all
setters correspond to getters, and that running all getters on a copied
object all return the same values) and relies on getting the method's
names via reflection, which is going to cause a lot of issues as
ProGuard usage increases.

BUG=620323

Review-Url: https://codereview.chromium.org/2186993003
Cr-Commit-Position: refs/heads/master@{#408397}