Ash window cycle UI - fix highlight painting during animation.
The background was being painted to the size of the view, which changes
immediately. Instead it should be painted to the size of the layer,
which transitions from one size to another.

Previously, when transitioning from highlighting a wider window to a narrower
one, it would resize immediately. When transitioning from highlighting a
narrower window to a wider one, it would appear to work correctly because the
background would be clipped to the layer bounds (which were animating) but
the rounded corners were subtly cut off.

BUG=626111

Review-Url: https://codereview.chromium.org/2185173002
Cr-Commit-Position: refs/heads/master@{#408418}