Fix order of operations in LayoutView::mapAncestorToLocal
mapAncestorToLocal should first recurse, then apply each of the offsets/transforms
in the forward direction but inverse order.

The forward direction will be transformed to backwards direction by the TransformState state bit indicating "unapply inverse direction".
This was incorrectly fixed in https://codereview.chromium.org/1994203003.

Also, remove code to handle transforms on the LayoutView. It's impossible
to have such a transform from content. This code was added in:

https://chromium.googlesource.com/chromium/src/+/ce75f4f9824faae4464420cf9c96a6cf0fb85c77

without comment as to why.

BUG=613894

Review-Url: https://codereview.chromium.org/2141693002
Cr-Commit-Position: refs/heads/master@{#408434}