Componentize spellcheck [2]: move common/ files to component.
Move chrome/common spellcheck files to component, update
spellcheck namespace, update dependencies and build files,
reformat.

The main motivation behind the componentization of the
spellcheck feature is that we want to make it available
in Android WebView.

Preceding 'componentize' patch:
[1] https://codereview.chromium.org/2166683003/

Note:
Skipping presubmit due to a warning about IPC_ENUM_TRAITS,
which is to be addressed separately.

BUG=583616, 629609
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2177343002
Cr-Commit-Position: refs/heads/master@{#408442}