CreateButtonBackground() has been creating unpremultiplied bitmaps.
This has been feeding unpremultiplied SkBitmaps into Skia as if they were premultiplied when drawing UI themes.  The attached bug shows an example of how this can go wrong.

It looks like this bug is ~7 years old.  While reading the code, I noted down a couple questions worth thinking about if we want to tread into this code again.

As it is, this should be enough to be able to revert the earlier bandaid CL I landed (see the bug).  It should also let us start asserting that premultiplied pixels are well-formed, to prevent this sort of problem in the future.

BUG=611002

Review-Url: https://codereview.chromium.org/2079413002
Cr-Commit-Position: refs/heads/master@{#408484}