Fix visible position calculation in accessible setSelection
The accessible setSelection function is similar to the equivalent APIs to
create a range and set the selection in the DOM, except that offsets are
all visible positions, since the accessibility tree uses visible text
(with whitespace collapsed, for example).

The code that created a VisiblePosition from a text node and offset was
incorrect if the text node had a previous sibling. Fix it so that the
offset of the text node within its container is taken into account.

BUG=626385
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2164813003
Cr-Commit-Position: refs/heads/master@{#408520}