Fail write_build_config.py if current_toolchain != default_toolchain
There are a few spots in the Android templates that assume appending
__build_config to a given target label will result in the label for a
write_build_config action. When toolchains exist in a label, appending
like this doesn't work. E.g.: :some_label(//toolchain)__build_config

Previously, we added a GN assert to enforce default toolchain, but that
was too strict, as it made it so no other targets that are defined in
the same BUILD.gn file could use a non-default toolchain.

This change make the error a build-time error rather than a gn gen
error.

BUG=629371

Review-Url: https://codereview.chromium.org/2161063003
Cr-Commit-Position: refs/heads/master@{#408576}