ImageCapture: add WebRtcImageCaptureBrowserTest content_browsertest
This CL adds a content_browsertest for ImageCapture:
WebRtcImageCaptureBrowserTest, with one test case,
CreateAndGetCapabilities, that is parameterized to
exercise the FakeVideoCaptureDevice and, in the very
near future, the system's device if implemented
(i.e., Android).

WebRtcImageCaptureBrowserTest needs access to the
physical webcam, but in general content_browsertests
are always using the Fake. WebRtcWebcamBrowserTest has
the necessary infrastructure to do this, so it's made
public and reused as base class for the former.

Also, this CL moves
 content/test/webrtc_content_browsertest_base.{cc,h}
to
 content/browser/media/webrtc/
since they are purely internal to content_browsertests.

BUG=627537, 518807
TEST=new content_browsertests
WebRtcImageCaptureBrowserTest.CreateAndGetCapabilities/0
 (all platforms)

Review-Url: https://codereview.chromium.org/2190523004
Cr-Commit-Position: refs/heads/master@{#408690}