[arc-intents] Properly qualify API-generated page transitions.
Unqualified page transitions are treated as link-clicks, which can get
sent to android via the arc intent mechanism.  This can result in
infinite-looping if the URL that is being opened originated from android
in the first place.

Marking this as an API-generated page transition ensures that the arc
intent code does not forward this URL back to android.

BUG=632360

Review-Url: https://codereview.chromium.org/2189063002
Cr-Commit-Position: refs/heads/master@{#408775}