[MD settings] only add overscroll if needed.
This CL changes the overscroll padding so that it will be empty (zero
height) if no overscroll is needed to bring the selection to the top of
the window.

This Cl does not fix the overscroll during window resizing or hide the
overscroll when scrolling up. Those will be in a future CL.

BUG=631332
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2201553002
Cr-Commit-Position: refs/heads/master@{#408847}