[TrackedPref] a super MAC validating a null contents is also valid.
This fixes the issue found @
https://codereview.chromium.org/2112713004/#msg39 (PS 14-15).

ProfilePrefStoreManagerTest.NewPrefWhenFirstProtecting is verifying
that the following works:
1) Have N valid MACs for tracked but so far unprotected prefs.
2) Start tracking a new pref and enable protection for all tracked prefs
3) Existing tracked prefs should be migrated.
4) The newly tracked pref should also be accepted per the old MAC being valid.

This failed because https://codereview.chromium.org/2112713004
correctly no longer creates the DictionaryValue representing the MACs
in DictionaryHashStoreContents when RemoveEntry() is called while it
doesn't even exist (i.e. nothing to remove).

But not creating the dictionary results in the end of the transaction stamping
a null contents instead of an empty contents which results in a different MAC
although it's essentially protecting the same thing.

The logic altered in this CL previously refrained from declaring a MAC stamping
a null contents as valid while it's just as valid and secure as one stamping an
empty dictionary.

BUG=624858

Review-Url: https://codereview.chromium.org/2198873002
Cr-Commit-Position: refs/heads/master@{#408964}