Use text affinity to return correct accessible line boundaries.
Our calculations of line boundaries for accessibility were wrong
in certain cases where the character offset is ambiguous. It's
possible for there to be a line break immediately between two
characters, such that one character is on the previous line and
one is on the next line. If the character index is between those
two characters, previously we had no way of knowing if the cursor
was on the first line or second line.

Fix this by passing through the text affinity of the selection
focus and anchor from Blink all the way into Chrome's accessibility
tree. If the affinity is upstream, it means the cursor is on the
previous line, and downstream means the next line.

I added a LayoutTest for Blink and a unit test for Chrome. I also
manually verified this fixes at least one broken case with VoiceOver,
see the bug for a very specific repro.

BUG=632450

Review-Url: https://codereview.chromium.org/2191833003
Cr-Commit-Position: refs/heads/master@{#408989}