Refactoring Me2MeNativeMessagingHost to implement NativeMessageHost interface
This change updates the Me2Me native messaging host to use the same interface
as the It2Me native messaging host.  This change reduces the differences
between the two hosts and will make it simpler to continue the refactoring in
my next change to allow Me2Me and It2Me native messaging hosts to share the
same process elevation logic and communication channel.

BUG=617185

Review-Url: https://codereview.chromium.org/2153313002
Cr-Commit-Position: refs/heads/master@{#409046}