Fix window drawing after snap on Windows 10.
Sometimes during window snap the window will receive
WM_WINDOWPOSCHANGING with a new size, but not WM_WINDOWPOSCHANGED. This
causes Chrome to avoid drawing to the window because it thinks that
might cause the window contents to be stretched.

To prevent this post a windows message in the WM_WINDOWPOSCHANGING
handler, and if that message is received without a WM_WINDOWPOSCHANGED
then re-enable drawing.

BUG=562681

Review-Url: https://codereview.chromium.org/1811913007
Cr-Commit-Position: refs/heads/master@{#409096}