RTCPeerConnection.getStats gets LegacyInterfaceTypeChecking.
This tag was removed in https://codereview.chromium.org/2156063002/
which caused a regression (crbug.com/632457). This CL re-adds it to
solve that regression. The generated bindings code is able to
call the correct overloaded getStats function with or without it.

LegacyInterfaceTypeChecking means any (nonsense) argument can be passed
as the second parameter and if its not a MediaStreamTrack it is
interpreted as null. This was something adapter.js relied on.

BUG=632457, 627816

Review-Url: https://codereview.chromium.org/2198123002
Cr-Commit-Position: refs/heads/master@{#409167}