Fix tests: conformance_textures_image_bitmap_from_image_data on Android
Right now these tests are failing, and the reason is that in our previous
CL here: https://codereview.chromium.org/2183323003, we made a change
such that the function swizzleImageData() is called only when
kN32_SkColorType == kBGRA_8888_SkColorType. That means on the Android
platform this function will not be called. What I didn't realize is that
this function itself also contains flip the input data vertically which
is required when flipY is true. Right now on the Android bots, whenever
the flipY is true, these tests will fail.

This CL changes the function swizzleImageData() so that this function
determines whether swizle is needed or not based on this:
kN32_SkColorType == kBGRA_8888_SkColorType

BUG=633214

Review-Url: https://codereview.chromium.org/2202143002
Cr-Commit-Position: refs/heads/master@{#409225}