Obsolete old PLT.* histograms in favor of PageLoad.* histograms.
In a subsequent change in a few months time, we'll remove
page_load_histograms.cc, which is responsible for logging these
histograms. We want to give a few months time where the histograms
are still logged, in case there are still users of these histograms
that we're not currently aware of.

See
https://groups.google.com/a/google.com/d/topic/chrome-metrics-team/icBD6tSvMyo/discussion
for more context.

BUG=595742

Review-Url: https://codereview.chromium.org/1837673003
Cr-Commit-Position: refs/heads/master@{#409272}