Remove SCT counters from DevTools security panel
This CL concerns the information that the DevTools security panel shows
for the Signed Certificate Timestamps (SCTs) that were served on a
request. (SCTs are part of the Certificate Transparency project.)

Each SCT has a validation status, and initially (in
https://codereview.chromium.org/1589703002), the security panel showed a
count of how many SCTs were served with each status.

Later, in https://codereview.chromium.org/1772603002, we added the full
details of each SCT to the security panel. Thus the counters are now
somewhat redundant: we show "X valid SCTs" followed by a summary of each
SCT with its validation status.

This CL removes the counters ("X valid SCTs, Y invalid SCTs, ...").

While the counters are a little more scannable at a glance, they
clutters the UI with redundant information and present an extra burden
for maintaining the plumbing needed to show the counters. This is
relevant right now because we want to add an additional SCT status. We
could rework the plumbing to accommodate this additional SCT status,
but it seems to make more sense to just remove the redundant information
from the UI.

BUG=591848,634006

Review-Url: https://codereview.chromium.org/2208803002
Cr-Commit-Position: refs/heads/master@{#409666}