bluetooth: Remove Web Bluetooth from Bad Flags Prompt.
Initially the --enable-web-bluetooth flag was considered risky because
the chooser UI had not yet been implemented -- thus not giving users
an affordance to accept or deny the request. The first device found
was always returned, premission presumed.

Now the chooser UI is implemented on all platforms, and
--enable-web-bluetooth no longer needs to be considered a bad flag.

This partially reverts
 "bluetooth: Add a WebBluetooth command line flag"
 https://codereview.chromium.org/1134473003

BUG=632106

Review-Url: https://codereview.chromium.org/2192573002
Cr-Commit-Position: refs/heads/master@{#409702}