Remove the swizzling of -[NSEvent window] and the middle button in EventsMacTest
The NSEvent initializers for creating test events are limited, so we
currently swizzle some properties. In a follow-up, we also want test
events that include phase information for scrolling. The swizzling for that
was becoming complex, but we can replace the swizzling by making more
use of CGEventSetIntegerValueField(..) and adding a tiny bit of reflection.

BUG=615948
TEST=events_unittests still passing

Review-Url: https://codereview.chromium.org/2209003002
Cr-Commit-Position: refs/heads/master@{#409718}