Add histograms for usage of WebSocket binary types
WebCore.WebSocket.BinaryTypeChangesAfterOpen counts the number of times
the value of binaryType was changed after the WebSocket was open, per
connection. No-op changes (eg. "blob" -> "blob") are not counted.

WebCore.WebSocket.ReceivedBlobMessageSize is a histogram of the size of
binary messages received when binaryType was "blob".

WebCore.WebSocket.ReceivedArrayBufferMessageSize is the same thing for
binary messages received when binaryType was "arraybuffer".

WebCore.WebSocket.SendBlobMessageSize is a histogram of sizes of
messages that were sent as Blobs.

WebCore.WebSocket.SendArrayBufferMessageSize is the same thing for
messages sent as ArrayBuffers.

WebCore.WebSocket.SendArrayBufferViewMessageSize is the same thing for
messages sent as ArrayBufferViews.

Known issue: BinaryTypeChangesAfterOpen fails to count WebSocket connections
that were not closed before the renderer was destroyed.

BUG=584616
TEST=manual inspection of about:histograms

Review-Url: https://codereview.chromium.org/1672433004
Cr-Commit-Position: refs/heads/master@{#409750}