cc: Add a flag to draw green borders on GL composited quads.
Debugging overlays can be tricky since when it works the output might
visually be the same as when it doesn't.

Adding DebugBorderQuads doesn't help since it interferes with overlay
strategies.

This CL adds a flag to draw green borders on top of texture quads
as soon as the texture quads has been draws using GL.
When a texture quad gets promoted to an overlay the border will
not be drawn.

BUG=b/29430506
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_precise_blink_rel

Review-Url: https://codereview.chromium.org/2198573002
Cr-Commit-Position: refs/heads/master@{#409889}