MacViews: Claim -[NSView isOpaque] for the content view according to the root layer properties.
Once we start scrolling with Layers, most of the CPU taken up when
scrolling is actually taken up by redrawing the window frame. This is
because BridgedContentView currently always claims that it is
transparent. So, when paint areas are invalidated, AppKit must first ask
the window frame to redraw that region in case BridgedContentView wants
to draw something transparent on top of that.

For opaque windows, none of the window background shows through, so
report YES from -[NSView isOpaque] in those cases.

BUG=594907, 605131

Review-Url: https://codereview.chromium.org/2219663003
Cr-Commit-Position: refs/heads/master@{#410012}