Format GN files in-place by default.
Almost all manual uses of "gn format" are to update files in-place. Make
this mode the default rather than requiring --in-place to be passed.

With this change, you can update a file in-place or use pipes, but the previous
mode to take input from a file and output to stdout is lost. This seems to be
not a strong use-case, and one can always pipe a file to stdin to work around
the limitation.

Suppresses outputting formated data when --dump-tree is specified.

Review-Url: https://codereview.chromium.org/2219703003
Cr-Commit-Position: refs/heads/master@{#410137}