Centralize replaced content rect computation to a single function
This CL re-purposed LayoutReplaced::replacedContentRect() for calculating
replaced content box of a replaced element. The original replacedContentRect()
was renamed to computeObjectFit() to reflect what it really does.

In addition to basic CSS box model, replaced elements have an extra box for
the actual placement of the replaced contents. Previously we didn't have
a centralized function for calculating this box and each LayoutReplaced
subclass implements in slightly different way. This could lead to duplicated
code and inconsistencies.

Review-Url: https://codereview.chromium.org/2199073002
Cr-Commit-Position: refs/heads/master@{#410167}