Fix CompositorWorker properties for scrolling and root layer.
Prior to this patch, the kScrollTop and kScrollLeft properties were inexplicably set mutable on every composited scrolling layer.

Also, this patch rationalizes the way scrolling-related mutable properties are applied to frame scrolls, in preparation for root layer scrolling.

When root layer scrolling is disabled, scrolling-related properties are applied to the root scrolling layer in PaintLayerCompositor::updateIfNeeded().

When root layer scrolling is enabled, PaintLayerCompositor::m_scrollLayer is obsolete, and the mutable properties for frame scrolls are applied the same way as properties on non-frame scrolls (i.e., in CompositedLayerMapping::updateElementIdAndCompositorMutableProperties for the LayoutView's PaintLayer).

There is, however, one hitch: because there is no Element associated with the LayoutView, updateElementIdAndCompositorMutableProperties must apply properties set on document.scrollingElement to the LayoutView's PaintLayer, hence the logic around document.scrollingElement() in this patch.

Also parameterize CompositorWorkerTest to run with and without root layer scrolling enabled, to prevent regressions.

BUG=490942
R=vollick@chromium.org,skobes@chromium.org

Review-Url: https://codereview.chromium.org/2147313002
Cr-Commit-Position: refs/heads/master@{#410200}