Task manager should support Idle Wakeups on Windows
This change adds a special SharedSampler that can refresh
multiple TaskManager resources for all chrome Task Groups
in a single operation.

On OS_WIN the implementation is based on
NtQuerySystemInformation and currently provides
only Idle Wakeups per second to start with. It should be
fairly straightforward to support a few other resource
types later, for example CPU usage and Private Memory.

On non-Windows platforms SharedSampler has a trivial stub
implementation that does nothing.

BUG=620832

Review-Url: https://codereview.chromium.org/2178733002
Cr-Commit-Position: refs/heads/master@{#410236}