[Android] Make download manager to support tablet
This CL does two things:
1. Introduce DownloadPage, which is the interface DownloadManagerUi
interact with the NativePage framework of Chrome.
2. Corresponding refactoring of the existing download manager code to
make it extendable. FilterAdapater/SpaceDisplay are spinned off from the
main class for better code structure.

One thing to note is that DownloadUiObserver is introduced as well,
which will become the way that the main manager talk to its sub
components.

BUG=616324
R=dfalcantara@chromium.org

Review-Url: https://codereview.chromium.org/2212183002
Cr-Commit-Position: refs/heads/master@{#410399}