Enable GPU acceleration in 2D OffscreenCanvases on main thread
This change allows OffscreenCanvasRenderingContext2D to create
an AcceleratedImageBufferSurface for its backing on platforms
that support GPU-accelerated canvas.

The plumbing of the runtime option for controlling the
availability of accelerated canvases was moved from WebSettings
to RuntimeEnableFeatures so that we no longer need to access
the DOM in order to access the feature flag.

BUG=593514

Review-Url: https://codereview.chromium.org/2192493003
Cr-Commit-Position: refs/heads/master@{#410423}