Mac: Fix PlatformFontMac::DeriveFont.
1. Currently, when we derive a font from an existing PlatformFontMac instance
   with only the size changed, we fail to persist the Font::UNDERLINE style in
   the derived font.

2. Also in all the other cases, instead of using the native_font_ instance to
   derive the font, we create a new PlatformFontMac instance from the font_name
   and other parameters. This can lead to unexpected results when the font from
   which we are deriving is a system font, since Apple explicitly forbids
   using a system font name to create a new font (See Potential API Pitfalls
   section in the WWDC talk Introducing the New System Fonts. Link -
   https://developer.apple.com/videos/play/wwdc2015/804/).

For example, in the Text Styles example of views_examples_with_content_exe, on
clicking underline, the derived font does not have underline due to 1 and is
bold because of 2, even though we didn't specify the bold font trait.

This CL modifies PlatformFontMac::DeriveFont to use the NSFontManager to create
a new derived NSfont. Since the new PlatformFontMac instance created has the
passed font style, 1 is solved. Also, since we no longer use the font name to
create the derived font, 2 is also solved. Also, added two tests which fail on
the current master and demonstrate the problem.

BUG=605404
TEST=Build and run views_examples_with_content_exe. Select the "Text Styles"
example. Check "Underline". Verify the text gets underlined. Verify the font is
not bold. Play around with different combinations and verify that they work as
expected.

Review-Url: https://codereview.chromium.org/2222483002
Cr-Commit-Position: refs/heads/master@{#410525}