[Media Router WebUI] Append new sinks to the end of the list.
Currently, when the sink list is updated in the WebUI, we just replace |sinksToShow_| in the container. This sometimes causes re-sorting and sinks suddenly appear in a different place. This can result in accidental casting to the wrong sink.

This change appends new sinks to the end of the sink list. A user who has incrementally discovered sinks will see the dialog grow vertically as new sinks are appended at the bottom of the list.

Also updated variable naming to make it more easily readable.

BUG=634365
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2221933003
Cr-Commit-Position: refs/heads/master@{#410585}