Don't commit a second blank document when going back to about:blank.
In the UseSubframeNavigationEntries codepath, loading subframe history
items is asynchronous.  This means the parent frame has a chance to
script an about:blank iframe before we hear the history item to load.
In the common case that we just go back to the about:blank page for
the iframe anyway, we do not want to clobber any injected contents by
loading a second about:blank in the frame.

This fix leaves one remaining bug, where we don't generate a commit
for the about:blank iframe.  For now, we just simulate DidStopLoading,
but we'll want to find a way to commit it instead.

BUG=626416
TEST=No error when restoring Google Docs tabs.
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/2222223002
Cr-Commit-Position: refs/heads/master@{#410748}