[WebApks] Tweak logic for 'bouncing back to chrome'
Previously a reload of a page inside a WebApk can erroneously redirect
back to Chrome (see bug for further details). Unfortunately, due to the
display logic of WebApks, we can't rely on startActivity doing the right
thing because Android can't tell that the MainActivity for the shell apk
is just a 'bouncer' to chrome's WebApkActivity. Add a necessary short-circuit.
BUG=604262

Review-Url: https://codereview.chromium.org/2225893004
Cr-Commit-Position: refs/heads/master@{#410788}