[Chromoting] Multitexture support for GlDesktop
This CL adds functionality to GlDesktop to break down the desktop frame
into multiple smaller textures if the host resolution exceeds the hardware
texture size limit. This is part of the project of using OpenGL to render
host on the Chromoting Android client.

This CL also refactors texture ID's for different objects.

BUG=385924

Review-Url: https://codereview.chromium.org/2211553002
Cr-Commit-Position: refs/heads/master@{#410848}