Created an intentionally crashing test for Android Instrumentation test.
It is important for us to be sure that we capture all the
right information if a testcase crashes, information such
as tombstones and logcat. Most existing testcases crash unpredictably,
so it is hard for us to check.

In this cl, I created an intentionally crashing testcase.
With this testcase, we are able to reliably check
how our test runner behaves when a testcase crashes.

BUG=

Review-Url: https://codereview.chromium.org/2216773002
Cr-Commit-Position: refs/heads/master@{#410919}