Enable ImeThread feature as default for M54
We are finch-experimenting this feature on M53. With dev experiment so
far, we could not find any problem. By enabling it by default on M54,
we will try to understand its impact on WebView better, and be ready to
launch this feature 100% on M54.

Also, do not crash even when registration fails since they are most
likely caused by view focus loss or window detachment, in which cases
we do not need to register anyways.

BUG=551193

Review-Url: https://codereview.chromium.org/2202853004
Cr-Commit-Position: refs/heads/master@{#411008}