Fix integration test failure on desktop.
The test provider for desktop just sends plain text message to client, but the
client expects it is a JSON string, which is mismatched. I don't know
why it worked before. But it is failed right now due to the message
format dismatch.

The fix is pretty straightforward, the test provider sends plain text
message to client and the client accepts plain text message as well.

Review-Url: https://codereview.chromium.org/2232623002
Cr-Commit-Position: refs/heads/master@{#411042}