X11: Only intersect _NET_WORKAREA with the primary display
GTK gives an explanation of this
https://github.com/GNOME/gtk/blob/952d0fd23fa9dbcf26aedf5bfd91c54a6addaf49/gdk/x11/gdkmonitor-x11.c#L69

"The EWMH constrains workarea to be a rectangle, so it can't adequately deal
with L-shaped monitor arrangements. As a workaround, we ignore the workarea for
anything but the primary monitor. Since that is where the 'desktop chrome'
usually lives, this works ok in practice."

BUG=510079

Review-Url: https://codereview.chromium.org/2235563002
Cr-Commit-Position: refs/heads/master@{#411124}