services/ui: Inject GpuService instance where needed, instead of singleton.
ui::GLES2Context uses the GpuService instance to establish a channel to the GPU
process. Instead of using the global singleton, pass through the GpuService
instance to ui::GLES2Context through the ui::OutputSurface. This removes the
last usage of the GpuService singleton, which means this code can be moved into
the client-lib, where it belongs.

BUG=634377
TBR=ben@ for change related to new API in content/renderer

Review-Url: https://codereview.chromium.org/2233003003
Cr-Commit-Position: refs/heads/master@{#411262}