Avoid an NullPointerException in android.util.Pair on Jelly Bean.
On some versions of Android Jelly Bean it appears that Pair.hashCode()
will encounter a NullPointerException if one of the elements in the pair
is null. This patch works around it by having WebsitePermissionsFetcher
store a non-null value as the second element that won't have any effect
on its operation.

BUG=636330

Review-Url: https://codereview.chromium.org/2231263002
Cr-Commit-Position: refs/heads/master@{#411346}