Update ToolbarActionViewInteractiveUITest to wait for menu closure
In a previous review (https://codereview.chromium.org/1489893003/) the app menu was switched to async. During this the toolbar interactice tests were updated as the menu no longer blocked returning to the tests.

This change restores the order of checking end conditions for two tests. It adds a wait to each, as the widget destruction associated with menu teardown is delayed on the message loop.

TEST=ToolbarActionViewInteractiveUITest.ActivateOverflowedToolbarActionWithKeyboard, ToolbarActionViewInteractiveUITest.TestClickingOnOverflowedAction
BUG=557134

Review-Url: https://codereview.chromium.org/2231423002
Cr-Commit-Position: refs/heads/master@{#411384}