Fix window/display bounds with multiple windows in mus.
Previously mus assumed that all displays and root windows had origin of
(0, 0) and were the primary display. This doesn't work with multiple
displays so use the proper display origin in screen space. Also keep
track of the primary display id.

This will ensure that root ServerWindows and ScreenMus have the correct
bounds for each display. This does not handle resizing or removing
displays correctly.

BUG=625876,614070

Review-Url: https://codereview.chromium.org/2230963003
Cr-Commit-Position: refs/heads/master@{#411385}