ImageCapture: only keep one set-zoom request in flight and catch IllegalStateException
This CL addresses a crash in ImageCapture when zoom is set
many times quickly in a row: |m_previewSession| was in some
borked state, causing onConfigured() to launch an
IllegalStateException, which was not caught.

Aside from not crashing, it's not necessary (and is slow)
to configure a zoom level if there is another request for
another zoom being queued, so this CL also changes the
behaviour to only have one zoom request in flight.

BUG=634985

Review-Url: https://codereview.chromium.org/2239503002
Cr-Commit-Position: refs/heads/master@{#411473}