Stop the complete() timer if the user closed UI.
If the user has closed the UI while the merchant is processing the
transaction, then the timer waiting for website to call complete()
should be stopped. Closing the UI disconnects the Mojo connection from
the renderer to the browser. When the timer finally fires, accessing the
Mojo connection causes a crash. This patch stops the timer when the Mojo
connection closes, thus preventing the crash.

BUG=637013

Review-Url: https://codereview.chromium.org/2244563002
Cr-Commit-Position: refs/heads/master@{#411697}