Combine Physical Web URLs into groups
A group id is an identifier that associates URLs into buckets.  (i.e.,
URLs from the same domain with the same title might have the same group
id).

This change displays groups as a single entry in our display list, and
then chooses the closest URL when a user taps the entry.

BUG=633715

Review-Url: https://codereview.chromium.org/2224223002
Cr-Commit-Position: refs/heads/master@{#411724}