Enable ash window cycle UI by default.
I removed the enable flag without replacing it with a disable flag
because I don't foresee a need for individuals to disable this feature.
In the case that we decide against launching, this patch should be
reverted (which is the same thing we'd have to do if we changed it to
a disable flag).

BUG=626111

Review-Url: https://codereview.chromium.org/2234233002
Cr-Commit-Position: refs/heads/master@{#411909}