aw: Workaround use after free due to finalizers
CleanupReference based on WeakReference can run before finalizers. So
the mere existance of a finalizer can keep enough webview objects alive,
causing use after free. The correct fix of using PhantomReference has
negative effects. So this is just a workaround.

The workaround is for apps that unintentionally uses a finalizer, but do
not intentionally try to use the webview in or past the finalizer. Just
mark the webview object as destroyed, but then silently swallow the
error to avoid crashes.

BUG=618807

Review-Url: https://codereview.chromium.org/2245713002
Cr-Commit-Position: refs/heads/master@{#411961}