Make AutomationNode.children about twice as fast.
From examining the performance characteristics of other bindings (next/previous sibling), this cl takes the same assumption that context switching in and out of V8 is expensive and should be minimized.

This cl adds a native binding to fetch all child ids in one request rather than making one request per child.

On a large site with 4000 elements, the time to fetch children with
root.children;

takes about
118ms without this cl and
70ms with this cl.

Subsequent requests to
root.children

are much faster
50ms without this cl

20ms with this cl.

Review-Url: https://codereview.chromium.org/2242163002
Cr-Commit-Position: refs/heads/master@{#412023}