navigator.share: Resolve promise only after user chooses a target.
Previously, would immediately resolve as soon as the share dialog was
shown. This introduces a new issue where the promise *never* resolves
or rejects if the user cancels the share picker (I'm not sure how to do
this or if it's even possible).

Introduces a bunch of new functionality in the ShareHelper class to
support this new use case.

Reland of https://crrev.com/410511 (to fix https://crbug.com/637194).

BUG=595608,637194

Review-Url: https://codereview.chromium.org/2203273002
Cr-Commit-Position: refs/heads/master@{#412175}