Add a feature to display a persistence toggle for permission prompts on Android.
This CL adds a disabled by default feature which shows a "Remember my
decision" toggle on permission prompts on Android. This is designed to
test whether a more explicit representation of permission persistence
(and better control over it) will encourage users to make decisions on
these prompts more often. It is intended to experiment with this
feature from M54.

A new PermissionInfoBar class is added to implement the optional
toggle. Permission-specific code in ConfirmInfoBar is moved into the
new class, which will also be beneficial in the future if permission
prompts stop being infobars. The permission decision is changed
from a boolean allowed/not allowed to the PermissionAction enum
already used for permission metrics reporting.

New metrics to measure the toggle position of prompts which are
granted or denied are also added in this CL to measure whether or
not users actually interact with the toggle (or if merely seeing it
is sufficient to change the wider permission action metrics).

The toggle is explicitly disabled for all permissions except
geolocation, which currently generates the most prompts on Android.

A subsequent CL will implement a corresponding checkbox for
desktop permission prompts using the same feature. The desktop
permission decision will also be changed to a PermissionAction.

BUG=632269

Review-Url: https://codereview.chromium.org/2226633002
Cr-Commit-Position: refs/heads/master@{#412177}