Remove the "card verified" message in PaymentRequest.
The merchant website has the final say in whether the provided credit
card is usable for the purchase. The merchant notifies the browser via
complete("fail") when the card is rejected. Chrome shows an error dialog
to the user when this happens. Therefore, CVC unmask dialog should not
show the "card verified" message. This is doubly true for local cards,
which are not verified against the server due to the privacy model. This
patch removes the "card verified" message from PaymentRequest. Autofill
card unmasking still has this message.

BUG=637454

Review-Url: https://codereview.chromium.org/2244893002
Cr-Commit-Position: refs/heads/master@{#412269}