Show the Cast toolbar icon ephemerally when Cast is in use
This CL changes the behavior of the Cast/Media Router toolbar component
action icon. Currently, the only way to add the icon is by installing
the Cast extension from the Web Store. After this change, the icon will
be displayed temporarily whenever the Media Router dialog is open, or there
is an active local media route.

This CL also replaces the "Remove icon" option in the action context menu
with "Always show icon" which can be checked. When this option is unchecked,
the icon is neither on the toolbar nor in the overflow menu, and is only shown
ephemerally in the situations mentioned above.

We replace the word "popup" with "dialog" wherever possible in
MediaRouterAction, to stay consistent with the rest of Media Router files.

We also change the default position of the component actions on the toolbar
from the leftmost to the rightmost (same behavior as extension actions).

TBR=grt@chromium.org
BUG=594577

Review-Url: https://codereview.chromium.org/2155293002
Cr-Commit-Position: refs/heads/master@{#412280}