[Extensions] Move Click-To-Script prefs to one file, handle bad values
Move all click-to-script preference handling to the
ScriptingPermissionsModifier.  Previously, the setting of the pref value
itself was done in extension_util (which is kind of a catch-all file
that should be avoided if possible), while the preference initialization
was done in ScriptingPermissionsModifier.  This CL moves them all to the
same place.

Additionally, add logic to handle bad preferences.  In some cases (e.g.
if the user installed an extension that was later migrated to a policy
extension), there could be a preference set for an extension being
allowed on all urls that's at-odds with the ability to withhold
permissions. For instance, if the user installs Extension A, withholds
all urls, and then Extension A becomes policy-installed, the preferences
for Extension A still indicate otherwise. Handle this case and migrate
the preferences to be correct for the policy-installed extension.

BUG=629927

Review-Url: https://codereview.chromium.org/2206903005
Cr-Commit-Position: refs/heads/master@{#412414}