Componentize spellcheck [4]: spellcheck/browser and android java-side.
Move some files and unit tests from chrome/browser/spellchecker/ and
android-specific java code to the spellcheck component. This move
generally includes the android slice of spellchecker.

In particular in this patch:
 - move the relevant chrome/browser/spellchecker files and unit tests
 - move the android java files and create proper jni registrar
 - isolate spellcheck preference constants into separate files
 - properly initialize "java_object_initialization_failed_" in
   spellchecker_session_bridge_android.cc
 - update references in the code base
 - update and create build files

The main motivation behind the componentization of the spellcheck
feature is that we want to make it available in Android WebView.

Preceding spellcheck componentization patches:
[1] https://codereview.chromium.org/2166683003/
[2] https://codereview.chromium.org/2177343002/
[3] https://codereview.chromium.org/2198143002/

BUG=583616, 629609
TBR=jam@chromium.org

Review-Url: https://codereview.chromium.org/2244083002
Cr-Commit-Position: refs/heads/master@{#412549}