Fix out of order key events in Chrome on Linux.
The ibus-gtk layer was using XPutBackEvent which pushes an event to the
top of the X event queue. The problem with this is that if the gtk event
loop processes more than one event before the X event loop executes the
message pump will see the events in reverse order.

Fix this by processing the mapped X11 event immediately. Since the gtk
event loop and the X11 event loop are on the same thread this makes
sense as the events stay in the proper order.

BUG=524084

Review-Url: https://codereview.chromium.org/1324913002
Cr-Commit-Position: refs/heads/master@{#412607}