Revert previous CL making PaintTabBackgroundUsingFillId() a non-method
BUG=621139

In preparation for https://codereview.chromium.org/2118853002/, two other
CLs were landed first.  Because the main CL was abandoned in favor of a
simpler approach, we are reverting one of these two preparatory CLs:

https://codereview.chromium.org/2197613002/ made major changes to the
prototype of PaintTabBackgroundUsingFillId() in order to force it to be
a non-method.  These and some related changes are now unnecessary, and
are being reverted here.  NOTE: The revert was done manually, so other
subsequent changes should be preserved.

(NOTE: https://codereview.chromium.org/2150013002/ moved a number of
static members and methods of the Tab class into the anonymous namespace,
along with a few methods that were "mostly static".  These changes are
considered a general code improvement, and are being left alone.)

Review-Url: https://codereview.chromium.org/2255823002
Cr-Commit-Position: refs/heads/master@{#412691}