Update WebAPK if homescreen icon changed.
This CL makes ManifestUpgradeDetector refetch the homescreen icon. The icon
specified in the "icons" property of the Web Manifest
https://www.w3.org/TR/appmanifest/ is fetched if specified. If the icon URL
changed or the icon bitmap data has changed an upgrade request is sent to
the WebAPK server.

If the "icons" property of the Web Manifest is empty, the homescreen icon URL
that the WebAPK is currently using is refetched to check whether it has changed.
If it has changed an upgrade request is sent to the WebAPK server. For the sake
of simplicity, the ManifestUpgradeDetector does not determine the page's favicon
and fetch the favicon if the "icons" property of the Web Manifest is empty.

BUG=624834
TEST=ManifestUpgradeDetectorTest.*

Review-Url: https://codereview.chromium.org/2206493002
Cr-Commit-Position: refs/heads/master@{#412713}