Attempt to rebaseline remaining crbug.com/463358 tests
There are 2 tests remaining marked for crbug.com/463358. They
both look like they need rebaselines, and maybe they are now fixed
and stable. There is still some chance that the Linux Debug
result for fast/transforms/transformed-caret.html is broken or
flakey, and will need to be re-added after the rebaseline.

No test should really be marked under this bug. It relates to
a clang roll that was itself reverted, so the cause of the failures
no longer exists.

TBR=fmalita@chromium.org
BUG=463358,636475

Review-Url: https://codereview.chromium.org/2252163002
Cr-Commit-Position: refs/heads/master@{#412881}