cc: Call GLRenderer::SwapBufferComplete only if a swap happened.
Display::DrawAndSwap used to call DirectRenderer::SwapBuffersComplete even
if DirectRenderer::SwapBuffers was not called.
This would cause the number of calls to DirectRenderer::SwapBuffers and
DirectRenderer::SwapBuffersComplete to mismatch.

GLRenderer assumes the number of SwapBuffersComplete calls matches the
number of SwapBuffer calls, and uses that assumption to release resources.

This CL avoids calling SwapBuffersComplete if SwapBuffers has not been
called so that the equilibrium between SwapBuffers and SwapBuffersComplete
is restored.

BUG=b/29430506
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_precise_blink_rel

Review-Url: https://codereview.chromium.org/2256673003
Cr-Commit-Position: refs/heads/master@{#412913}