Avoid creating sync token when only change is new hardware video frame.
Once we get the picture ready IPC the GPU has already done all the
processing it needs to for the frame, so the renderer doesn't need to
create a new sync token. We can instead use the same sync token every
frame to avoid a synchronous IPC to the GPU process.

BUG=575856

Review-Url: https://codereview.chromium.org/2254993003
Cr-Commit-Position: refs/heads/master@{#412934}