Import compare_build_artifacts.py from isolate recipe module.
Build determinism is necessary for improving try performance.
If two test binaries are the same, we usually do not need to run the test.

The imported script checks the two builds are the same and shows
error if they are different.

The script is imported to make it easy for chromium developers to
update the script and blacklist json to keep the script updated.

BUG=314403

Review-Url: https://codereview.chromium.org/2232723002
Cr-Commit-Position: refs/heads/master@{#413016}