MD History: Improve test helpers, update synced tabs tests to use them
Many of our existing tests rely heavily on the state of the whole page.
While this is necessary in many cases, we should avoid it when possible
and take steps to ensure the page state is properly reset between tests.

This CL adds new test helpers to support completely resetting the page
for each individual test. As part of this, CSS mixin/variable
definitions are moved so that they apply once to <html> rather than to
many different elements.

Finally, the test suite for <history-synced-device-manager> is updated
to use the new helpers. Most tests now directly test the manager
element, with no <history-app> integration. A separate test suite for
integration tests with a <history-app> is also available.

BUG=425625
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2257703002
Cr-Commit-Position: refs/heads/master@{#413033}