Unify CSSVariableResolver::resolveVariableReferences functions
There are currently two CSSVariableResolver::resolveVariableReferences
functions, which differ only by the first argument, one taking a
StyleVariableData* and one taking a StyleResolverState&. The logic in
these is written slightly differently but they are actually identical
(the latter could just call the former with state.style().variables()).

This patch unifies the two functions, keeping a function with a
const StyleResolverState& argument as future work will need more data
from the state object and not just the variables object.

Review-Url: https://codereview.chromium.org/2261633002
Cr-Commit-Position: refs/heads/master@{#413147}