Provide a callback for Component Updater OnDemand calls.
The idea here is that there is no good way for a client to know
when a subsequent OnDemand call can be issued without getting
an Error::ERROR_UPDATE_IN_PROGRESS. This issue showed up when writing a
browser test for an unrelated component updater feature.

This change is mostly mechanical, especially for the existing users
of the OnDemand interface. All existing call sites pass a null callback
as an argument, therefore no new code paths are introduced with this change.

BUG=639180

Review-Url: https://codereview.chromium.org/2261533003
Cr-Commit-Position: refs/heads/master@{#413178}