Adjust WebsiteSettings statuses for subresources with cert errors
SSLPolicy and SecurityStateModel used to record subresources with
certificate errors as mixed content. As of
https://codereview.chromium.org/2226363002/, however, subresources with
cert errors are recorderded separately, in their own
|content_with_cert_errors_status| field on
SecurityStateModel::SecurityInfo.

This CL updates WebsiteSettings to warn the user when there are
subresources with cert errors on the page. Currently, WebsiteSettings
uses the same statuses as are used for mixed content, because the strings happen
to apply well to both types of insecure content. (I also renamed a few strings and
enums to make it more clear that they describe different types of insecure
subresources, not just mixed content.) In future, maybe we'd want to use different
strings to describe subresources with cert errors in WebsiteSettings,
but I think the mixed content strings work well enough for now.

(Other UI like DevTools will distinguish mixed content from subresources with
cert errors, which is one of the reasons that we separated the two into
|mixed_content_status| and |content_with_cert_errors_status|.)

For simplicity, WebsiteSettings ignores subresources with certificate
errors when the main resource was loaded with a certificate error. I
think that trying to distinguish subresource cert errors from cert
errors on the main resource is probably TMI for the average user.

BUG=634171

Review-Url: https://codereview.chromium.org/2244243002
Cr-Commit-Position: refs/heads/master@{#413190}