bluetooth: Set rssi as valid in Chrome OS for linux
Chrome OS for linux uses a fake state machine that continuously updates
a device's rssi. The state machine never actually sets the "rssi" property
as valid so when clients try to read the rssi we always return that the rssi
is not available. This change sets the rssi property as valid so when
the rssi is updated clients can get a value for the property.

Review-Url: https://codereview.chromium.org/2264443003
Cr-Commit-Position: refs/heads/master@{#413203}