[sql] Retry post-poison open as soon as possible.
sql::Connection::Open() retries the open if the handle is poisoned by
the error callback.  The assumption is that the error callback likely
razed or recovered the database, and since there can be no outstanding
statements the open can be safely retried.  This change adds this
handling to call which attempts to probe the database for basic
validity.

BUG=none

Review-Url: https://codereview.chromium.org/2258703004
Cr-Commit-Position: refs/heads/master@{#413292}