[Android] Add user actions for bookmark manager creation
When checking the recent data about bookmark opening stats, we found a
mystery that is hard to explain: the number of bookmark open is larger
than the number of bookmark manager open on phones. This CL adds some
extra loggings to the bookmark code to seek definite answer to this
mystery.

BUG=639512

Review-Url: https://codereview.chromium.org/2266633002
Cr-Commit-Position: refs/heads/master@{#413456}