Change zoom presets to be more useful.
Device pixel ratio at 1.25 and 1.5 device scale factors are problematic
for existing zoom levels. In particular there was no 80% zoom level that
allowed users with 1.25 set to achieve a dpr of 1 (i.e. equivalent to
no zoom or dsf). Also 1/3 and 2/3 were set to 0.333 and 0.666 instead of
their more precise representations, which caused the math to be slightly
off when using those (especially important for 1.5 dsf, because 2/3 is
needed to achieve a dpr of 1). This patch fixes both of these issues.

R=dbeam@chromium.org
BUG=632323
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2263733002
Cr-Commit-Position: refs/heads/master@{#413494}