Recreate the WebData database on a catastrophic SQL error
Currently, if the WebData database gets corrupted, it will remain like so
and users will keep getting the profile error dialog every time they
sign in. The only way for the user to fix this is to recreate the profile
or delete the corrupt file.
This CL does for the WebData DB as is done for the History DB; when a catastrophic
error is seen, the DB is razed.

Example feedback report filed with this CL applied: https://docs.google.com/a/google.com/document/d/1qt-AwWVHZTtcJoFPNq2_yI-Tl2pIiOwEQCaIf23iL4g/edit?usp=sharing

BUG=455749, 635734

Review-Url: https://codereview.chromium.org/2225333003
Cr-Commit-Position: refs/heads/master@{#413500}