Scale Rect Origins Using the Specified HWND's Display
Previously, the origin and rect sizing were scaled independently, which
meant that the origin could be scaled using one display and the size
could be scaled with another display.

With this change, the origin uses the same display used with the size.

Bonus Fix: GetScreenWinDisplayNearestDIPRect now returns the correct
ScreenWinDisplay. If the display touched the DIP rect, the function
would immediately return when a intersecting display would have been
a better choice.

BUG=636521

Review-Url: https://codereview.chromium.org/2267123002
Cr-Commit-Position: refs/heads/master@{#413782}