Prevent synchronous layout while painting the last frame of a gif
This patch is a followup to [1] and removes another synchronous layout
in the animated gif painting code.

If an animated gif is behind, BitmapImage::startAnimation will advance
the animation until it catches up. This process is called "skipping
frames" and it should not cause a synchronous layout because it will
occur during paint. This patch fixes a bug when skipping frames hit
the last frame of a non-looping gif. Before this patch, we would
synchronously notify observers when the last frame was hit which would
end up causing a layout. With this patch, we now defer the
last-frame-needs-paint notification until paint is done.

An enum has been added for the skippingFrames mode which may make it
easier to understand.

[1] https://crrev.com/45eef1a2a0f721e6782c4c92b28de3835f94ec46

BUG=623798

Review-Url: https://codereview.chromium.org/2262493003
Cr-Commit-Position: refs/heads/master@{#413784}