[MD User Menu] New way to handle supervised user auth error
Background:
The MD user menu shows an error header when sync errors occur. For auth
errors in particular, the header prompts the user to sign in again to
resolve them. However, when the user is a supervised user, signing in is
not an option.

In this CL, we add a way to specifically handle auth errors for
supervised users, which displays a different message in the error header
and with no action button.

Screenshot: https://drive.google.com/a/google.com/file/d/0B7Fvv7JszRyGaUtrbUNXVDBlR0k/view

BUG=615893

Review-Url: https://codereview.chromium.org/2267013002
Cr-Commit-Position: refs/heads/master@{#413847}