Separate PaymentRequest initialization and display.
In order to better accommodate delays in querying payment apps, separate
initialization of PaymentRequest from displaying it. Thus a merchant
website can create a JavaScript PaymentRequest object on page load, but
call PaymentRequest.show() only when user clicks the [BUY] button on the
page. With this patch, Chrome begins to query the installed payment apps
before the user clicks the [BUY] button. This enables faster startup of
the PaymentRequest UI.

BUG=638454

Review-Url: https://codereview.chromium.org/2250793003
Cr-Commit-Position: refs/heads/master@{#413893}