WebSocket: Stop Bridge::disconnect() from waiting until Peer::disconnect() is complete
Bridge::disconnect() on a worker thread waited until Peer::disconnect() is
complete on the main thread to make sure that Peer never dereferences a pointer
to Bridge after disconnection. However, in the current implementation, the
pointer is CrossThreadWeakPersistent and its existence is appropriately checked
before access[1], so Bridge::disconnect() no longer have to wait until it's
nullified on Peer::disconnect().

[1] https://codereview.chromium.org/2123703002

BUG=639153, 640088

Review-Url: https://codereview.chromium.org/2271133002
Cr-Commit-Position: refs/heads/master@{#414040}