build: Refactor get_concurrent_links.py
Instead of passing build config arguments to the python
script, allow caller to provide the memory needed per linker
and the amount of memory to reserve. Then move the logic
specific to build config to gn file, which is a better place
in case the logic gets more complicated in the future.

Then clean up the script. Factor out a function to get the
total ram of system, and unify the logic for calculating the
number of concurrent links.

Note this intends to be a no-op change, but refactored code
is not identical, and there may be corner case differences.

BUG=617429

Review-Url: https://codereview.chromium.org/2259503002
Cr-Commit-Position: refs/heads/master@{#414084}