Remove anchor view from container when reset
Makes sure ScopedAnchorView::Reset() removes the anchor view
from its container view through JNI. This prevents potential
leaks when the destructor fails to do the cleaning job. Not
apparently related to the bug, but it helps avoid any issues
that may happen in the future.

Also deleted a method declaration of |SetWindowAndroid| sneaked
in by mistake.

BUG=635515

Review-Url: https://codereview.chromium.org/2272673003
Cr-Commit-Position: refs/heads/master@{#414154}