Fix document mode migration.
There is an old tab_state file in the tabbed mode state directory. If
it is not deleted and the new tab state is written to the new
tab_state0 file, then when it comes time to run the migration code in
TabPersistentStore the old tab_state file will be given preference
and overwrite the new one with all the document mode tabs that were
just migrated. Thus we need to delete the old tab_state file if it
exists.

BUG=640625

Review-Url: https://codereview.chromium.org/2274943002
Cr-Commit-Position: refs/heads/master@{#414194}