Safeguard against crash-looping if subresource filter rule indexing fails.
A sentinel file is placed in the ruleset version directory just before indexing
commences, and removed afterwards. Therefore, if a sentinel file is found on
next start-up, it is an indication that the previous indexing operation may
have crashed, and indexing will not be attempted again.

This CL also slightly refactors unit tests so that we no longer have specific
tests verifying histograms values, but instead we have tests exercising error
scenarios, and verifying histogram values is performed as part of that.

TBR=rkaplow@chromium.org
BUG=609747

Review-Url: https://codereview.chromium.org/2272593002
Cr-Commit-Position: refs/heads/master@{#414235}