Port context_lost test to new GpuIntegrationTest harness.
The main trick here is to get hand-written tests to trampoline through
the _RunGpuTest harness, which is where test expectations are
implemented.

This is done by adopting a convention that the "real" test, like
"_GPUProcessCrashesExactlyOnce", is prefixed with an underscore. This
is implemented in self-contained fashion in
ContextLostIntegrationTest.GenerateGpuTests. If necessary, this will
be generalized further into a base class.

The waterfalls will be switched to run the new version of the test in
a follow-on CL.

BUG=352807
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_optional_gpu_tests_rel;master.tryserver.chromium.mac:mac_optional_gpu_tests_rel;master.tryserver.chromium.win:win_optional_gpu_tests_rel

Review-Url: https://codereview.chromium.org/2271543002
Cr-Commit-Position: refs/heads/master@{#414279}