arc: Stop piggybacking "text input started" and "show virtual keyboard" events.
It is not only useless in the latest version, but is even harmful
for some advanced apps that take the liberty of splitting these
two events.

This change itself should not change any observable behavior yet,
but will be necessary to enable coming Android-side change to
make this split effective.

BUG=640816
BUG=b/31001453
TEST=ArcImeServiceTest
TEST=Tested manually with some ARC apps (Settings, PlayStore)

Review-Url: https://codereview.chromium.org/2280503002
Cr-Commit-Position: refs/heads/master@{#414283}