Remove stale comments for blink::decodeURLEscapeSequences()
There are several comments and FIXMEs about concern about NULL
unescaping.

The FIXMEs have been there since the commit for upstreaming
KURLGoogle.cpp which is
http://crrev.com/1730ee062f4e55666b1bbf3e777443e508c60708.

But the commit
http://crrev.com/8e70a3c3b73ed85dd62102bf251683feda91307e has already
removed the special care for NULL character. The internal bug b/1350291
was just mentioning WebKit bug 20559.

We still have a warning comment about brettw@'s concern in KURL.h, so,
clean up the others.

As a bonus, this CL adds a test to check that the method decodes UTF-8.

R=brettw
BUG=none

Review-Url: https://codereview.chromium.org/2270373002
Cr-Commit-Position: refs/heads/master@{#414286}