Renderer-side changes for NoState Prefetch
NoState Prefetch design doc:
https://docs.google.com/document/d/16VCYGGWau483IMSxODpg5faZny1FJ6vNK2v-BuM5EhU

This CL is the first step for NoState Prefetch implementation
in the renderer.

The code added in this CL is not called yet (it will be called
only when the PREFETCH_ONLY enum is sent through IPC).

The prefetching is implemented in the HTMLDocumentParser.
When prefetching, the parser runs in single-thread mode.
When feeded with HTML input, it runs the preload scanner on
it, and stops without actually building or executing the page.

BUG=632361

Review-Url: https://codereview.chromium.org/2172613002
Cr-Commit-Position: refs/heads/master@{#414406}