Re-enable program memory cache for Adreno 4/5xx
The program cache was disabled on Adreno 4/5xx due to crashes.

I'm hoping that this issue was caused by invalidations/issues with the
on-disk program cache, possibly with driver updates or other issues.

To test this, I'm re-enabling the in-memory portion of the cache
only. The cached objects will not outlive the GPU process.

I've also confirmed that the related Adreno 3xx bug, crbug.com/510673,
does not impact the Adreno 400 (and I'm guessing 500, but will track
down a device to double check).

We should watch crash rates for gfx::GLApiBase::glProgramBinaryFn to
make sure this change doesn't re-introduce these crashes.

BUG=598060, 486117
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_precise_blink_rel

Review-Url: https://codereview.chromium.org/2267163003
Cr-Commit-Position: refs/heads/master@{#414490}