VTVDA: Fix computation of |reorder_window|.
Prior to this change, VTVDA would disable reordering when there was no
VUI max_num_reorder_frames syntax element. This is incorrect; excluding
an explicit list of profiles, it should be MaxDpbFrames.

The incorrect behavior is masked during normal playback, because
decoding is typically faster than rendering. This causes the reorder
queue to fill up even though reordering is supposed to be turned off,
and frames therefore get a chance to be ordered correctly.

BUG=634007

Review-Url: https://codereview.chromium.org/2272233002
Cr-Commit-Position: refs/heads/master@{#414499}