Fix NFC crasher from accessing a destroyed activity.
While this works, it would be better to have a clearer signal that
the Activity died (or is in the process of shutting down) so we
could make a better timed decision to disable NFC reading.

Also note that Activity#isDestroyed() is only available in
JELLY_BEAN_MR1, so this is not the best solution for all device
implementations that need to determine this.

BUG=640890

Review URL: https://codereview.chromium.org/2303523002 .

Review-Url: https://codereview.chromium.org/2292353002
Cr-Original-Commit-Position: refs/heads/master@{#415441}
Cr-Commit-Position: refs/branch-heads/2840@{#88}
Cr-Branched-From: 1ae106dbab4bddd85132d5b75c670794311f4c57-refs/heads/master@{#414607}