[Temp] Only depend on devices/vr when enable_webvr is on.
This is a backport of some of the build system bits of
http://crrev.com/2219203002 ("Migrate WebVR Cardboard implementation to
GVR"): we are not interested in the "move to GVR" part, but this change
also makes the dependency on some devices/vr targets optional based on
the value of `enable_webvr`, which is something we want in order to be
able to build Crosswalk with that flag off.

This commit will not be necessary once we move to M54 and incorporate
the CL mentioned above.

BUG=XWALK-6746