Restore replacedWillBeRemoved to fix compile after postMessage changes
It is still in use on the M54 branch (2840) but wasn't on master.

This can't be landed on master and backported, precisely because it
is unused, which causes a warning, and -Werror is enabled.

BUG=425896,640825
TBR=sigbjornf@opera.com,haraken@chromium.org
NOTRY=true
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2331293004
Cr-Commit-Position: refs/branch-heads/2840@{#325}
Cr-Branched-From: 1ae106dbab4bddd85132d5b75c670794311f4c57-refs/heads/master@{#414607}