Attempt to repair corrupt enterprise policy force-installed extensions
For normally installed extensions, when the content verification system
detects corruption we just disable the extension and provide a "Repair"
button on the chrome://extensions page that users can click to manually
reinstall the extension. But for enterprise force-installed extensions
that are supposed to always remain enabled, we weren't sure how to
handle this (since we have had a few bugs over time which resulted in
false-positive corruption detection in various edge cases, and this can
happen as a result of minor disk corruption as well as malicious
changes). So we just let them keep running but recorded an UMA histogram
to measure how often this happens. Looking at the data for this, it
happens very rarely - around 0.5% of total enterprise policy extensions
seem to be affected.

In some recent conversations with Google's own internal IT folks, we
decided we'd like to attempt to automatically repair instead of just
doing nothing, so this CL implements that. It also includes new UMA
stats for measuring the number of repair attempts and successes, as well
as issuing a warning to the browser log which enterprise admins could
have automated systems watching for. Over time we'd like to implement a
separate log facility for these kinds of events that are of interest
just to enterprise admins so that they don't have to turn on the general
browser log.

BUG=447040

Review-Url: https://codereview.chromium.org/2299203004
Cr-Commit-Position: refs/heads/master@{#417608}
(cherry picked from commit 56282ab78b4c684e7008f777a81924d823177817)

Review URL: https://codereview.chromium.org/2342483002 .

Cr-Commit-Position: refs/branch-heads/2840@{#346}
Cr-Branched-From: 1ae106dbab4bddd85132d5b75c670794311f4c57-refs/heads/master@{#414607}