Ensure CCTs have the proper context menu on tab restore.
Custom tab context menus worked previously because CustomTabActivity
was in charge of creating the initial tab, and all subsequent tabs
reuse the delegate of the parent tab.

When restoring, the TabModelSelectorImpl via TabPersistentStore
was creating the tabs and using the default logic in ChromeTabCreator.

This change moves ChromeTabCreator creation to the activities
themselves, which allows CCTs to overwrite the default tab delegate
construction.  This also moves the svelte tab saving logic to the
selector, which is consistent with all other tab saving behavior (and
allows the activity to be unaware of that dependency and enables
creation w/ fewer params).

BUG=649139

Review URL: https://codereview.chromium.org/2368683003 .

Review-Url: https://codereview.chromium.org/2359923005
Cr-Original-Commit-Position: refs/heads/master@{#420702}
Cr-Commit-Position: refs/branch-heads/2840@{#516}
Cr-Branched-From: 1ae106dbab4bddd85132d5b75c670794311f4c57-refs/heads/master@{#414607}